% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSatelliteUtils.R
\docType{data}
\name{bbox_CONUS}
\alias{bbox_CONUS}
\title{Bounding box for the 48 contiguous United States}
\format{
A vector with 4 numerics.
}
\usage{
bbox_CONUS
}
\description{
Shorthand for \code{c(-125, -65, 24, 50)}
}
\details{
bbox_CONUS
}
\keyword{datasets}
