% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSatelliteUtils.R
\docType{data}
\name{example_nc}
\alias{example_nc}
\title{Example NC}
\format{An NCDF4 dataset}
\description{
This dataset holds a small sample of satellite data from November 15th, 2018.
}
\keyword{datasets}
