% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{getAttributes}
\alias{getAttributes}
\title{Returns attributes from a netCDF file}
\usage{
getAttributes(nc, varid = 0, print = TRUE)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating 
what file to read from.}

\item{varid}{The variable whose attribute is to be read. Can be a character 
string with the variable's name or an object of class ncvar4. As a special 
case, if \code{varid == 0}, then a global (file) attribute will be read 
rather than a particular variable's attribute.}

\item{print}{Logical specifying whether human readable output should be
printed to the console.}
}
\value{
A named list of netCDF attributes is returned invisibly.
}
\description{
Returns attributes from a netCDF files
}
