% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSatelliteUtils.R
\name{getSatelliteDataDir}
\alias{getSatelliteDataDir}
\title{Get Satellite Data Directory}
\usage{
getSatelliteDataDir()
}
\value{
Absolute path string.
}
\description{
Returns the data directory where Satellite data is located.
}
\seealso{
\link{SatelliteDataDir}

\link{setSatelliteDataDir}
}
\keyword{environment}
