% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSatelliteUtils.R
\name{getValue}
\alias{getValue}
\title{Get raster value at a point}
\usage{
getValue(raster, x, y)
}
\arguments{
\item{raster}{raster layer}

\item{x}{x (longitude) value}

\item{y}{y (latitude) value}
}
\value{
value of raster cell covering the point
}
\description{
Get raster value at a point
}
\keyword{utils}
