% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_findClosestScanFile.R
\name{goasaodc_findClosestScanFile}
\alias{goasaodc_findClosestScanFile}
\title{Find the scan file closest to a datetime}
\usage{
goasaodc_findClosestScanFile(
  satID = NULL,
  datetime = NULL,
  timezone = NULL,
  baseUrl = "https://tools-1.airfire.org/Satellite/"
)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Datetime in any Ymd HMS [MS] format or \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime}.}

\item{baseUrl}{URL of remote database; Defaults to 
"https://tools-1.airfire.org/Satellite/".}
}
\value{
A scan filename.
}
\description{
Finds the name of the scan file that was generated closest to 
the given datetime.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

closestScanFile <- goasaodc_findClosestScanFile(
  satID = "G17",
  datetime = "2020-09-08 08:30:00",
  timezone = "America/Los_Angeles"
)

# Print precise time of scan
closestScanFile \%>\%
  goesaodc_convertFilenameToDatetime() \%>\%                          # UTC
  MazamaCoreUtils::parseDatetime(timezone = "America/Los_Angeles")  # Local
}
}
