% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goes_createSpatialPoints.R
\name{goes_createSpatialPoints}
\alias{goes_createSpatialPoints}
\title{Create a SpatialPointsDataFrame of GOES data}
\usage{
goes_createSpatialPoints(nc, xmn = NULL, xmx = NULL, ymn = NULL,
  ymx = NULL, dqfLevel = NULL)
}
\arguments{
\item{nc}{ncdf4 handle}

\item{xmn}{lower longitude extent}

\item{xmx}{upper longitude extent}

\item{ymn}{lower latitude extent}

\item{ymx}{upper latitude extent}

\item{dqfLevel}{data quality flag level}
}
\value{
SpatialPointsDataFrame
}
\description{
Create a SpatialPointsDataFrame of GOES data including data
within the specified extent and data quality flag level. Data quality level
can take a value of:

0: High quality retrieval flag
1: Medium quality retrieval flag
2: Low quality retrieval flag
3: No retrieval quality flag
}
