% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goes.R
\name{goes_getCoordGrid}
\alias{goes_getCoordGrid}
\title{Creates a data frame of decimal degree lon/lat pairs for a given
netcdf handle}
\usage{
goes_getCoordGrid(nc)
}
\arguments{
\item{nc}{netcdf handle}
}
\value{
Dataframe.
}
\description{
Creates a data frame of decimal degree lon/lat pairs for a given
netcdf handle
}
