% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goes.R
\name{goes_lonLat}
\alias{goes_lonLat}
\title{Converts GOES xy radian coordinates to latitude and longitude pairs}
\usage{
goes_lonLat(x, y, r_eq, r_pol, H, lambda0)
}
\arguments{
\item{x}{x value from the netcdf in radians (scale and offset already applied)}

\item{y}{y value from the netcdf in radians (scale and offset already applied)}

\item{r_eq}{semi_major_axis}

\item{r_pol}{semi_minor_axis}

\item{H}{satellite height from center of earth = perspective_point_height +
semi_major_axis}

\item{lambda0}{longitude_of_projection_origin (converted to radians)}
}
\value{
a named list with lon and lat values
}
\description{
Converts GOES xy radian coordinates to latitude and longitude pairs
}
\examples{
\dontrun{
coords <- goes_lonLat(x_rad, y_rad, r_eq, r_pol, H, lambda0)
}
}
