% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_aggregateScanPoints.R
\name{goesaodc_aggregateScanPoints}
\alias{goesaodc_aggregateScanPoints}
\title{Aggregate GOES scan points}
\usage{
goesaodc_aggregateScanPoints(spdfList = NULL, fun = mean, na.rm = FALSE)
}
\arguments{
\item{spdfList}{A \code{list} of GOES scan \code{SpatialPointsDataFrame}s.}

\item{fun}{The function to use for aggregating AOD values. Defaults to 
\code{mean}.}

\item{na.rm}{Logical flag whether to remove NA values before performing the
\code{fun} function. Defaults to \code{FALSE}.}
}
\description{
Creates a \code{SpatialPointsDataFrame} of aggregated AOD 
readings from a \code{list} of GOES scan \code{SpatialPointsDataFrame}s.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanPointsList <- goesaodc_createScanPoints(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon
)

goesaodc_aggregateScanPoints(
  spdfList = scanPointsList,
  na.rm = TRUE
)
}
}
