% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_aggregateScanRasters.R
\name{goesaodc_aggregateScanRasters}
\alias{goesaodc_aggregateScanRasters}
\title{Aggregate GOES scan rasters}
\usage{
goesaodc_aggregateScanRasters(rasterBrick = NULL, fun = mean, na.rm = FALSE)
}
\arguments{
\item{rasterBrick}{A \code{RasterBrick} of GOES scan \code{RasterLayers}s.}

\item{fun}{The function to use for aggregating AOD values. Defaults to 
\code{mean}.}

\item{na.rm}{Logical flag whether to remove NA values before performing the
\code{fun} function. Defaults to \code{FALSE}.}
}
\description{
Creates a \code{RasterLayer} of aggregated AOD readings from a 
\code{RasterBrick} of GOES scan \code{RasterLayers}s.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

# Create a raster from scans averaged over a time range
scanRasterBrick <- goesaodc_createScanRaster(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon,
  cellSize = 0.05
)

goesaodc_aggregateScanRasters(
  rasterBrick = scanRasterBrick,
  na.rm = TRUE
)
}
}
