% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_calcAverageScanPoints.R
\name{goesaodc_calcAverageScanPoints}
\alias{goesaodc_calcAverageScanPoints}
\title{Calculate the average of a series of GOES scan points}
\usage{
goesaodc_calcAverageScanPoints(spdfList = NULL, na.rm = FALSE)
}
\arguments{
\item{spdfList}{A \code{list} of GOES scan \code{SpatialPointsDataFrame}s.}

\item{na.rm}{Logical flag whether to remove NA values before calculating the
average. Defaults to \code{FALSE}.}
}
\description{
Creates a \code{SpatialPointsDataFrame} of averaged AOD 
readings from a \code{list} of GOES scan \code{SpatialPointsDataFrame}s.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanPointsList <- goesaodc_createScanPoints(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon
)

goesaodc_calcAverageScanPoints(
  spdfList = scanPointsList,
  na.rm = FALSE
)
}
}
