% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_calcAverageScanRaster.R
\name{goesaodc_calcAverageScanRaster}
\alias{goesaodc_calcAverageScanRaster}
\title{Calculate the average raster of a series of GOES scan rasters}
\usage{
goesaodc_calcAverageScanRaster(rasterBrick = NULL, na.rm = FALSE)
}
\arguments{
\item{rasterBrick}{\code{RasterBrick} of GOES scan \code{RasterLayers}s.}

\item{na.rm}{Logical flag whether to remove NA values before calculating the 
average. Defaults to \code{FALSE}.}
}
\value{
\code{RasterLayer} that is the average of all the 
\code{RasterLayers} in a \code{RasterBrick}.
}
\description{
Creates a \code{RasterLayer} of averaged AOD readings from a 
\code{RasterBrick} of GOES scan \code{RasterLayers}s.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanFiles <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles"
)

# Create a raster from scans averaged over a time range
scanRasterBrick <- goesaodc_createScanRaster(
  filename = scanFiles,
  bbox = bboxOregon,
  cellSize = 0.05
)

goesaodc_calcAverageScanRaster(
  rasterBrick = scanRasterBrick,
  na.rm = TRUE
)
}
}
