% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_calcAverageScanSpdf.R
\name{goesaodc_calcAverageScanSpdf}
\alias{goesaodc_calcAverageScanSpdf}
\title{Calculate the average SPDF of a series of GOES scan SPDFs}
\usage{
goesaodc_calcAverageScanSpdf(spdfList = NULL, na.rm = FALSE)
}
\arguments{
\item{spdfList}{\code{list} of GOES scan \code{SpatialPointsDataFrame}s.}

\item{na.rm}{Logical flag determining whether to remove NA values before 
calculating the average. Defaults to \code{FALSE}.}
}
\value{
\code{SpatialPointsDataFrame} that is the average of all the
\code{SpatialPointsDataFrames} in a \code{list}.
}
\description{
Creates a \code{SpatialPointsDataFrame} of averaged AOD 
readings from a \code{list} of GOES scan \code{SpatialPointsDataFrame}s.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanFiles <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles"
)

spdfList <- goesaodc_createScanSpdf(
  filename = scanFiles,
  bbox = bboxOregon
)

goesaodc_calcAverageScanSpdf(
  spdfList = spdfList,
  na.rm = TRUE
)
}
}
