% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_calcTrendScanPoints.R
\name{goesaodc_calcTrendScanPoints}
\alias{goesaodc_calcTrendScanPoints}
\title{Calculate the trend of a series of GOES scan points}
\usage{
goesaodc_calcTrendScanPoints(spdfList = NULL, na.rm = FALSE)
}
\arguments{
\item{spdfList}{A \code{list} of GOES scan \code{SpatialPointsDataFrame}s.}

\item{na.rm}{Logical flag whether to remove NA values before calculating the
trend. Defaults to \code{FALSE}.}
}
\description{
Creates a \code{SpatialPointsDataFrame} of AOD reading trends 
from a \code{list} of GOES scan \code{SpatialPointsDataFrame}s. A trend value
for a point is calculated by taking the difference between it's average value
in the first half of the scan series and it's average value in the second.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanFiles <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles"
)

scanPointsList <- goesaodc_createScanPoints(
  filename = scanFiles,
  bbox = bboxOregon
)

goesaodc_calcTrendScanPoints(
  spdfList = scanPointsList,
  na.rm = TRUE
)
}
}
