% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_calcTrendScanRasters.R
\name{goesaodc_calcTrendScanRasters}
\alias{goesaodc_calcTrendScanRasters}
\title{Calculate the trend of a series of GOES scan rasters}
\usage{
goesaodc_calcTrendScanRasters(rasterBrick = NULL, na.rm = FALSE)
}
\arguments{
\item{rasterBrick}{A \code{RasterBrick} of GOES scan \code{RasterLayers}s.}

\item{na.rm}{Logical flag whether to remove NA values before calculating the
trend. Defaults to \code{FALSE}.}
}
\description{
Creates a \code{SpatialPointsDataFrame} of AOD reading trends 
from a GOES scan \code{rasterBrick}. A trend value for a cell is calculated 
by taking the difference between it's average value in the first half of the 
scan series and it's average value in the second.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanFiles <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles"
)

scanRasterBrick <- goesaodc_createScanRaster(
  filename = scanFiles,
  bbox = bboxOregon,
  cellSize = 0.05
)

goesaodc_calcTrendScanRasters(
  rasterBrick = scanRasterBrick,
  na.rm = TRUE
)
}
}
