% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_calcTrendScanSpdf.R
\name{goesaodc_calcTrendScanSpdf}
\alias{goesaodc_calcTrendScanSpdf}
\title{Calculate the trend SPDF of a series of GOES scan SPDFs}
\usage{
goesaodc_calcTrendScanSpdf(spdfList = NULL, na.rm = FALSE)
}
\arguments{
\item{spdfList}{A \code{list} of GOES scan \code{SpatialPointsDataFrame}s.}

\item{na.rm}{Logical flag determining whether to remove NA values before 
calculating the trend. Defaults to \code{FALSE}.}
}
\value{
A \code{SpatialPointsDataFrame} that is the trend of all the 
\code{SpatialPointsDataFrames} in a \code{list}.
}
\description{
Creates a \code{SpatialPointsDataFrame} of AOD trend values 
from a \code{list} of GOES scan \code{SpatialPointsDataFrame}s. A trend value
for a point is calculated by taking the difference between it's average value
in the first half of the scan series and it's average value in the second 
half.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanFiles <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles"
)

spdfList <- goesaodc_createScanSpdf(
  filename = scanFiles,
  bbox = bboxOregon
)

goesaodc_calcTrendScanSpdf(
  spdfList = spdfList,
  na.rm = TRUE
)
}
}
