% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_convertFilenameToDatetime}
\alias{goesaodc_convertFilenameToDatetime}
\title{Get the scan start time from a GOES AODC netCDF file name}
\usage{
goesaodc_convertFilenameToDatetime(file = NULL)
}
\arguments{
\item{file}{File name of GOES AOD file.}
}
\value{
The scan start time.
}
\description{
Gets the scan start time from a GOES AODC netCDF file name.
}
