% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createEmptyScanPoints.R
\name{goesaodc_createEmptyScanPoints}
\alias{goesaodc_createEmptyScanPoints}
\title{Create NA AOD points from a GOES scan}
\usage{
goesaodc_createEmptyScanPoints(
  satID = NULL,
  filename = NULL,
  bbox = bbox_CONUS
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest. Defaults to CONUS.}
}
\description{
Creates a \code{SpatialPointsDataFrame} of NA AOD readings from 
a GOES scan,
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

MazamaSatelliteUtils:::goesaodc_createEmptyScanPoints(satID = "G16")
MazamaSatelliteUtils:::goesaodc_createEmptyScanPoints(satID = "G17")
MazamaSatelliteUtils:::goesaodc_createEmptyScanPoints(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202530031174_e20202530033547_c20202530035523.nc"
)
}
}
