% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createHourlyRasterStack.R
\name{goesaodc_createHourlyRasterStack}
\alias{goesaodc_createHourlyRasterStack}
\title{Create a RasterStack for a specified hour}
\usage{
goesaodc_createHourlyRasterStack(satId = NULL, startdate = NULL,
  var = "AOD", res = 0.1, bbox = NULL, dqfLevel = NULL)
}
\arguments{
\item{satId}{ID number of the source GOES satellite}

\item{startdate}{startdate, specified to the hour, in any Y-m-d H format or
\code{POSIXct}}

\item{var}{variable ("AOD, "DQF" or "ID"). Set to AOD by default.}

\item{res}{resolution of raster in degrees. Set to 0.1 by default.}

\item{bbox}{bounding box generated by \code{sp::bbox}, \code{raster::extent} 
or a vector of floats in c(lonLo, lonHi, latLo, latHi) order.}

\item{dqfLevel}{data quality flag level, numeric (0, 1, 2, or 3)}
}
\value{
RasterStack
}
\description{
Create a \code{RasterStack} from GOES AOD data files for the 
date and hour specified by \code{startdate}. Each \code{RasterLayer} contains
data from one Advanced Baseline Imager (ABI) scan during the specified time 
period.

If data for the specified time period is not found in the directory specified 
by \code{setSatelliteDataDir()}, it will be downloaded in order to create the
\code{RasterStack}.

The Z axis of the \code{RasterStack} is a character vector where each element
is the time stamp of the scan and has the format YYYYMMDDHHMMSS. This can be 
accessed using the \code{raster::getZ()} function. Names of the 
\code{RasterStack} are also time stamps of the scan, of the format XHH.MM.SS.

Data quality level `dqfLevel` can take a value of:

0: High quality retrieval flag
1: Medium quality retrieval flag
2: Low quality retrieval flag
3: No retrieval quality flag
}
\examples{
\donttest{
### EXTENTS BASED ON PASSED IN FLOATS
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bbox_us <- c(-124.848974, -66.885444, 24.396308, 49.384358) # US LOWER 48

rstrStack <- goesaodc_createHourlyRasterStack(
satId = "G16", 
startdate = "2019-09-06 16", 
bbox = bbox_us,
dqfLevel = 2,
res = 0.2)

rasterVis::levelplot(rstrStack)

#### EXTENTS BASED ON sp::bbox OF OREGON
library(MazamaSpatialUtils)

setSpatialDataDir("~/Data/Spatial")
loadSpatialData("USCensusStates")

oregon <- subset(USCensusStates, stateCode == "OR")
bbox_oregon <- sp::bbox(oregon)

rstrStack <- goesaodc_createHourlyRasterStack(
satId = "G17", 
startdate = "2019-09-06 16", 
bbox = bbox_oregon,
dqfLevel = 2)

rasterVis::levelplot(rstrStack)
}
}
