% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createMultiScanPoints.R
\name{goesaodc_createMultiScanPoints}
\alias{goesaodc_createMultiScanPoints}
\title{Create spatial points for multiple scans}
\usage{
goesaodc_createMultiScanPoints(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  bbox = bbox_CONUS,
  dqfLevel = 3
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}
}
\description{
Creates a SpatialPointsDataFrame of averaged AOD values across
multiple scans.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("NaturalEarthAdm1")

bboxOregon <- c(-125, -116, 42, 47)

# Create points from scans covering a full hour
goesaodc_createMultiScanPoints(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon,
  dqfLevel = 2
)
}
}
