% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createMultiScanSPDF.R
\name{goesaodc_createMultiScanSPDF}
\alias{goesaodc_createMultiScanSPDF}
\title{Create spatial points for multiple scan}
\usage{
goesaodc_createMultiScanSPDF(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  bbox = bbox_CONUS,
  dqfLevel = 3
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}
}
\description{
Creates a SpatialPointsDataFrame of averaged AOD values across
multiple scans.
}
