% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createNativeGrid.R
\name{goesaodc_createNativeGrid}
\alias{goesaodc_createNativeGrid}
\title{Creates list of tibbles from a list of .nc files}
\usage{
goesaodc_createNativeGrid(nc_files = NULL, bbox = NULL)
}
\arguments{
\item{bbox}{geographic extents to limit data ingestion from NetCDF files to}

\item{ncFiles}{ncdf4 handle}
}
\value{
list of Tibbles (dataframe) with NetCDF variables and associated 
locations. List names correspond to timestamps of files in input list.
}
\description{
Create a list of tibbles with columns: AOD, DQF, lon and lat.
This information is sufficient to plot as points or create a raster object.
}
\examples{
\donttest{
# Process full extent of Gridfile
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadAOD(
  satID = "G16", 
  datetime = "201930019",
  endtime = "2019300211", 
  timezone = "UTC", 
  isJulian = TRUE)

my_files <- c("OR_ABI-L2-AODC-M6_G16_s20193001901344_e20193001904117_c20193001907158.nc")

tbl_list <- goesaodc_createNativeGrid(my_files)
head(tbl_list)

# Tibble based on BBOX filtered extent of tibble
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

my_files <- c("OR_ABI-L2-AODC-M6_G16_s20193001901344_e20193001904117_c20193001907158.nc",
              "OR_ABI-L2-AODC-M6_G17_s20193002116196_e20193002118569_c20193002121014.nc")

kincade_bbox <- c(-124, -120, 36, 39)

filtered_list <- goesaodc_createNativeGrid(
  my_files,
  bbox = kincade_bbox)
  
head(filtered_list)

}
}
