% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createRaster.R
\name{goesaodc_createRaster}
\alias{goesaodc_createRaster}
\title{Create a RasterBrick of GOES data}
\usage{
goesaodc_createRaster(
  nc = NULL,
  res = 0.1,
  fun = mean,
  bbox = bbox_CONUS,
  dqfLevel = NULL
)
}
\arguments{
\item{nc}{ncdf4 handle}

\item{res}{resolution of raster in degrees; Defaults to 0.1.}

\item{fun}{function to use when rasterizing; Defaults to mean.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Data quality flag level.}
}
\value{
RasterBrick
}
\description{
Create a RasterBrick of GOES AOD data including data points
with the specified resolution and function, and within the specified bounding
box and data quality flag level.

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or 
\code{raster::extent()}.

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

files <- goesaodc_downloadAOD(
  satID = "G17", 
  datetime = 2019102714, 
  timezone = "America/Los_Angeles",
  verbose = TRUE 
)

netCDF <- files[1]
nc <- goesaodc_openFile(netCDF)

# Kincade fire region
kincade_bbox <- c(-124, -120, 36, 39)

raster <- goesaodc_createRaster(
  nc, 
  res = 0.03,
  bbox = kincade_bbox,
  dqfLevel = 3
)
  
west <- c("washington", "oregon", "california", "nevada", "idaho")
maps::map(database = "state", regions = west)   
raster::plot(raster, "AOD", col=rev(heat.colors(5)), add = TRUE)
maps::map(database = "state", regions = west, add = TRUE)
}
}
