% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createScanPoints.R
\name{goesaodc_createScanPoints}
\alias{goesaodc_createScanPoints}
\title{Create spatial points from GOES scans}
\usage{
goesaodc_createScanPoints(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}
}
\description{
Creates a SpatialPointsDataFrame of AOD readings from one or 
more GOES scans. Scans are specified either by filename or satellite + time 
information.
}
\details{
If \code{filename} is given, just points for that scan file will be 
processed. Otherwise, scan(s) must be specified by \code{satID}, 
\code{datetime}, and \code{timezone}, which will determine the closest scan 
taken to that time. An average SPDF of a series of scans can be requested by 
also supplying \code{endtime}. All scans taken from \code{datetime} 
(inclusive) up to \code{endtime} (exclusive) will be processed by taking the 
average value of each point.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("NaturalEarthAdm1")

bboxOregon <- c(-125, -116, 42, 47)

# Create points for a scan specified by satellite + time
goesaodc_createScanPoints(
  satID = "G17",
  datetime = "2020-09-08 17:30",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon,
  dqfLevel = 3
)

# Create points for a scan file
goesaodc_createScanPoints(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202530031174_e20202530033547_c20202530035523.nc",
  bbox = bboxOregon,
  dqfLevel = 2
)

# Create points from scans averaged over a time range
goesaodc_createScanPoints(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon
)

# Create points for a faulty scan
goesaodc_createScanPoints(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
  bbox = bboxOregon
)
}
}
