% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createScanRaster.R
\name{goesaodc_createScanRaster}
\alias{goesaodc_createScanRaster}
\title{Create Rasters from GOES scans}
\usage{
goesaodc_createScanRaster(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3,
  cellSize = NULL,
  fun = mean
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Data quality flag level; Defaults to NULL.}

\item{cellSize}{Size of Raster cells measured in degrees lon/lat.}

\item{fun}{Function used to summarize multiple point values within a single 
cell; Defaults to \code{mean}.}
}
\description{
Creates a Raster of AOD readings from one or more GOES scans. 
Scans are specified either by filename or by satellite + time information.
}
\details{
If \code{filename} is given, just data for that scan file will be 
processed. Otherwise, scan(s) must be specified by \code{satID}, 
\code{datetime}, and \code{timezone}, which will determine the closest scan 
taken to that time. An average Raster of a series of scans can be requested 
by also supplying \code{endtime}. All scans taken from \code{datetime} 
(inclusive) up to \code{endtime} (exclusive) will be processed by taking the 
average value of each reading point and rasterizing the result (not by 
rasterizing each scan and then taking the average).
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("NaturalEarthAdm1")

bboxOregon <- c(-125, -116, 42, 46.5)

# Create a raster for a scan specified by satellite + time
goesaodc_createScanRaster(
  satID = "G17",
  datetime = "2020-09-08 17:30",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon,
  cellSize = 0.05
)

# Create a raster for a scan file
oregonRaster <- goesaodc_createScanRaster(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202530031174_e20202530033547_c20202530035523.nc",
  bbox = bboxOregon,
  dqfLevel = 2,
  cellSize = 0.05
)

# Create a raster from scans averaged over a time range
goesaodc_createScanRaster(
  satID = "G17",
  datetime = "2020-09-08 12:00",
 endtime = "2020-09-08 13:00",
 timezone = "America/Los_Angeles",
  bbox = bbox_oregon,
  cellSize = 0.05
)

# Create a raster for a faulty scan
goesaodc_createScanRaster(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
  bbox = bbox_oregon,
  dqfLevel = 3,
  cellSize = 0.05
)
}
}
