% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createScanSPDF.R
\name{goesaodc_createScanSPDF}
\alias{goesaodc_createScanSPDF}
\title{Create spatial points from GOES scans}
\usage{
goesaodc_createScanSPDF(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = NULL
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Data quality flag level; Defaults to NULL.}
}
\description{
Creates a SpatialPointsDataFrame of AOD readings from one or 
more GOES scans. Scans are specified either by filename or satellite + time 
information.
}
\details{
If \code{filename} is given, just points for that scan file will be 
processed. Otherwise, scan(s) must be specified by \code{satID}, 
\code{datetime}, and \code{timezone}, which will determine the closest scan 
taken to that time. An average SPDF of a series of scans can be requested by 
also supplying \code{endtime}. All scans taken from \code{datetime} 
(inclusive) up to \code{endtime} (exclusive) will be processed by taking the 
average value of each point.
}
