% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createTibble.R
\name{goesaodc_createScanTibble}
\alias{goesaodc_createScanTibble}
\title{Create a tibble from a ncdf4 handle}
\usage{
goesaodc_createScanTibble(nc = NULL, bbox = bbox_CONUS, dqfLevel = 3)
}
\arguments{
\item{nc}{ncdf4 handle or a list of handles.}

\item{bbox}{Bounding box for the region of interest. All points outside of 
this area will be removed. Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}
}
\value{
\code{tibble} with netCDF variables and associated locations.
}
\description{
Creates a tibble from a netCDF file with columns: AOD, DQF, lon 
and lat.
}
\examples{
\donttest{
# Tibble based on full extent of Gridfile
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

scanFile <- goesaodc_downloadScanFiles(
  satID = "G17", 
  datetime = "2019-10-27 10:00",
  timezone = "America/Los_Angeles" 
)

nc <- goesaodc_openScanFile(scanFile)

tbl <- goesaodc_createScanTibble(nc)
head(tbl)

# Tibble based on bbox filtered region

# Kincade fire region
bbox <- c(-124, -120, 36, 39)

filtered_tbl <- goesaodc_createScanTibble(nc, bbox)
head(filtered_tbl)
}
}
