% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createSingleScanPoints.R
\name{goesaodc_createSingleScanPoints}
\alias{goesaodc_createSingleScanPoints}
\title{Create spatial points for a single scan}
\usage{
goesaodc_createSingleScanPoints(
  satID = NULL,
  datetime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{The name of the scan file.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}
}
\description{
Creates a SpatialPointsDataFrame from a single scan file 
specified either by a filename or through satellite and time information.
}
\details{
If \code{filename} is given, \code{satID}, \code{datetime}, and
\code{timezone} will be ignored.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("NaturalEarthAdm1")

bboxOregon <- c(-125, -116, 42, 47)

# Create points from a scan specified by satellite and time
goesaodc_createSingleScanPoints(
  satID = "G17",
  datetime = "2020-09-08 17:30",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon
)

# Create points from a scan file
goesaodc_createSingleScanPoints(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202530031174_e20202530033547_c20202530035523.nc",
  bbox = bboxOregon,
  dqfLevel = 2
)
}
}
