% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createScanSpdf.R
\name{goesaodc_createSingleScanSpdf}
\alias{goesaodc_createSingleScanSpdf}
\title{Create an SPDF from a single GOES AOD scan}
\usage{
goesaodc_createSingleScanSpdf(filename = NULL, bbox = bbox_CONUS, dqfLevel = 3)
}
\arguments{
\item{filename}{Name of the scan file.}

\item{bbox}{Bounding box for the region of interest. Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}
}
\description{
Creates a \code{SpatialPointsDataFrame} of AOD readings from a 
GOES scan file.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanFile <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:30",
  timezone = "America/Los_Angeles"
)

MazamaSatelliteUtils:::goesaodc_createSingleScanSpdf(
  filename = scanFile,
  bbox = bboxOregon
)
}
}
