% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_downloadDaytimeAOD.R
\name{goesaodc_downloadDaytimeAOD}
\alias{goesaodc_downloadDaytimeAOD}
\title{Download GOES-16 or GOES-17 AOD data for the daylight hours of a 
specific date and location.

#' @description Download all daylight files available to directory specified 
with \code{setSatelliteDir}.  See \code{getDaylightHours()} for details 
regarding how daylight hours are calculated.}
\usage{
goesaodc_downloadDaytimeAOD(satID = NULL, datetime = NULL,
  longitude = NULL, latitude = NULL, bbox = NULL, timezone = NULL,
  isJulian = FALSE, verbose = FALSE)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Specific date for which daylight hours are requested.}

\item{longitude}{Longitude of the location of interest in decimal degrees E.}

\item{latitude}{Latitude of the location of interest in decimal degrees N.}

\item{bbox}{Bounding box for the region of interest, Default: CONUS.}

\item{timezone}{Timezone in which to interpret the \code{datetime}.}

\item{isJulian}{Logical value determining whether datetime should be 
interpreted as a Julian date with day of year as a decimal number.}

\item{verbose}{Logical specifies whether to show status of file download.}
}
\value{
Vector of filenames.
}
\description{
Download GOES-16 or GOES-17 AOD data for the daylight hours of a 
specific date and location.

#' @description Download all daylight files available to directory specified 
with \code{setSatelliteDir}.  See \code{getDaylightHours()} for details 
regarding how daylight hours are calculated.
}
\examples{
\donttest{

# List locally available files for a specific date by timezone

library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

goesaodc_downloadDaytimeAOD(satID = "G16",
datetime = "2019-09-06", 
timezone = "America/Los_Angeles")
}

}
