% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_downloadScanFiles.R
\name{goesaodc_downloadScanFiles}
\alias{goesaodc_downloadScanFiles}
\title{Download GOES AOD scan files}
\usage{
goesaodc_downloadScanFiles(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  isJulian = FALSE,
  filenames = NULL,
  verbose = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/"
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime as a Ymd HMS or Julian formatted string, or a 
\code{POSIXct}.}

\item{endtime}{End time as a Ymd HMS or Julian formatted string, or a 
\code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}. Defaults to UTC.}

\item{isJulian}{Logical flag determining whether \code{datetime} (and
optionally \code{endtime}) should be interpreted as a Julian date with day of
year as a decimal number. Defaults to FALSE.}

\item{filenames}{Names of scan files.}

\item{verbose}{Logical flag determining whether to print download progress 
messages.}

\item{baseUrl}{URL of remote database. Defaults to 
"https://tools-1.airfire.org/Satellite/".}
}
\description{
Downloads GOES AOD scan files for a given datetime or 
time range. If just \code{datetime} is given, then only the scan file closest
to that time will be downloaded. If \code{endtime} is specified as well, then
all scans from \code{datetime} up to (but not including) \code{endtime} will 
be downloaded.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:30",
  timezone = "UTC"
)

goesaodc_downloadScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "UTC",
  verbose = TRUE
)
}
}
