% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_getCoordBounds}
\alias{goesaodc_getCoordBounds}
\title{Determines the x and y bounds (in radians) for a given netcdf handle}
\usage{
goesaodc_getCoordBounds(nc)
}
\arguments{
\item{nc}{ncdf4 handle.}
}
\value{
A named list with x1, x2, y1, and y2.
}
\description{
Determines the x and y bounds (in radians) for a given netcdf handle
}
