% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_getCoordGrid}
\alias{goesaodc_getCoordGrid}
\title{Creates a data frame of decimal degree lon/lat pairs for a given
netcdf handle}
\usage{
goesaodc_getCoordGrid(nc)
}
\arguments{
\item{nc}{ncdf4 handle.}
}
\value{
Dataframe
}
\description{
Code borrowed from 
https://github.com/raffscallion/goesfire/blob/master/R/utils.R
}
