% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_getDaylightHours.R
\name{goesaodc_getDaylightHours}
\alias{goesaodc_getDaylightHours}
\title{Calculates the sunrise and sunset hours for a specific date and 
location}
\usage{
goesaodc_getDaylightHours(datetime = NULL, timezone = NULL,
  bbox = NULL, latitude = NULL, longitude = NULL, julian = FALSE)
}
\arguments{
\item{datetime}{specific date for which daylight hours are requested}

\item{timezone}{timezone in which to interpret the \code{datetime}}

\item{bbox}{Bounding box for the region of interest, Default: CONUS}

\item{latitude}{geographic latitude, in decimal degrees, for point of 
interest}

\item{longitude}{geographic longitude, in decimal degrees, for point of 
interest}

\item{julian}{logical, optional format to specify \code{datetime} parameter}
}
\value{
day_info List which contains sunrise and sunset times as POSIXt 
values
}
\description{
Returns the sunrise and sunset hours for a specific date and 
location based on sun ephemerides calculations.  Location can be specified 
via any of the following methods.

\itemize{
\item{}{Olson timezone, either as part of a POSIXt datetime, or specified 
via the \code{timezone} parameter}
\item{}{Geographic Bounding Box, either as a vector of floats in 
\code{c(lonLo, lonHi, latLo,latHi)} order, or the return value from 
\code{sp::bbox()}, or \code{raster::extent()} }
\item{}{Longitude and Latitude specified in decimal degrees}
}
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

# EX 1. Get Sunrise and Sunset by Lat/Lon and Datetime
dayInfo <- goesaodc_getDaylightHours(
datetime = "2019-09-06", 
longitude = -123.245, 
latitude = 42.861)

# EX 2. Get Sunrise and Sunset by BBOX and datetime
dayInfo <- goesaodc_getDaylightHours(
datetime = "2019-09-06",
bbox = c(-124.566, -116.463, 41.991, 46.292)) # Oregon

# EX 3. Get Sunrise and Sunset by Datetime and Timezone
dayInfo <- goesaodc_getDaylightHours(
datetime = "2019-09-06",
timezone = "America/Los_Angeles")
}

}
