% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_getStartString}
\alias{goesaodc_getStartString}
\title{Get the scan start time string from a GOES AODC netCDF file name}
\usage{
goesaodc_getStartString(file)
}
\arguments{
\item{file}{GOES AOD netCDF file name}
}
\value{
The scan start time string in Julian days.
}
\description{
Get the scan start time string from a GOES AODC netCDF file name
}
