% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_getStartTime}
\alias{goesaodc_getStartTime}
\title{get the scan start time from a GOES AOD netCDF file name}
\usage{
goesaodc_getStartTime(file)
}
\arguments{
\item{file}{GOES AOD netCDF file name}
}
\value{
scan start time
}
\description{
get the scan start time from a GOES AOD netCDF file name
}
