% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_installGoesGrids.R
\name{goesaodc_installGoesGrids}
\alias{goesaodc_installGoesGrids}
\title{Install GOES grids}
\usage{
goesaodc_installGoesGrids()
}
\value{
Invisibly returns a vector of local GOES 16 and 17 grid files.
\code{datetime}.
}
\description{
Creates data files with geolocation information for GOES-16 
(East) and GOES-17 (West) satellite products.  Checks for the presence of
GOES 16 & GOES 17 .nc (NetCDF) files in the directory previously set by 
\code{setSatelliteDataDir}. If present, will read projection and 
coordinate grid information from them and create GOES East and West grids in 
that directory. If .nc files are missing, will download appropriate ones 
and use them to create the grids.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

goesaodc_installGoesGrids()
} 
}
\seealso{
\code{\link{goesaodc_createGoesGrid}}
}
