% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_isGoesProjection}
\alias{goesaodc_isGoesProjection}
\title{Check if `nc` projection is GOES projection}
\usage{
goesaodc_isGoesProjection(nc = NULL)
}
\arguments{
\item{nc}{ncdf4 handle.}
}
\value{
logical
}
\description{
Determines if projection information in `nc` matches GOES 
projection information.
}
