% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listDaytimeFiles.R
\name{goesaodc_listDaytimeFiles}
\alias{goesaodc_listDaytimeFiles}
\title{List GOES AOD files available for a specific date and 
location during daytime hours}
\usage{
goesaodc_listDaytimeFiles(
  satID = NULL,
  datetime = NULL,
  longitude = NULL,
  latitude = NULL,
  bbox = NULL,
  useRemote = FALSE,
  timezone = NULL,
  isJulian = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/"
)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Datetime in any Ymd H [MS] format or \code{POSIXct}.}

\item{longitude}{Longitude of the location of interest in decimal degrees E.}

\item{latitude}{Latitude of the location of interest in decimal degrees N.}

\item{bbox}{Bounding box for the region of interest.}

\item{useRemote}{Logical specifying whether to look for files in 
\code{getSatelliteDataDir()} or \code{baseUrl}; Defaults to FALSE;}

\item{timezone}{Timezone used to interpret \code{datetime}.}

\item{isJulian}{Logical value determining whether \code{datetime} should be 
interpreted as a Julian date with day of year as a decimal number; Defaults 
to FALSE.}

\item{baseUrl}{URL of remote database; Defaults to 
"https://tools-1.airfire.org/Satellite/".}
}
\value{
Vector of filenames.
}
\description{
Retrieves a list of GOES AOD files for the daylight hours of a 
given \code{datetime}, available either locally in the specified 
\code{setSatelliteDir} or on a remote server using the \code{useRemote} 
parameter. See \code{getDaylightHours()} for details regarding how daylight 
hours are calculated.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

# List local files available for specific date by timezone
goesaodc_listDaytimeFiles(
  satID = "G16",
  datetime = "2019-09-06", 
  timezone = "America/Los_Angeles"
)

# List remote files available for specific date by timezone
goesaodc_listDaytimeFiles(
  satID = "G16", 
  datetime = "2019-09-06", 
  timezone = "America/Los_Angeles",
  useRemote = TRUE
)
}
}
