% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listFiles.R
\name{goesaodc_listFiles}
\alias{goesaodc_listFiles}
\title{List downloaded GOES AOD files for a specified date and hour}
\usage{
goesaodc_listFiles(date, hour = NULL)
}
\arguments{
\item{date}{desired date (integer, characer representing YYYYMMDD)}

\item{hour}{desired hour (integer, character representing HH)}
}
\value{
Vector of matching files
}
\description{
Retrieve a list of GOES AOD files available in the
satelliteDataDir for a specified date and hour
}
