% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listLocalFiles.R
\name{goesaodc_listLocalFiles}
\alias{goesaodc_listLocalFiles}
\title{List downloaded GOES AOD files for a specified date and hour}
\usage{
goesaodc_listLocalFiles(satID = NULL, datetime = NULL,
  timezone = "UTC", bbox = c(-125, -65, 24, 50), fullDay = FALSE,
  daylightOnly = TRUE)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Desired datetime in any Ymd H [MS] format or \code{POSIXct}}

\item{timezone}{Timezone in which to interpret the \code{datetime}.}

\item{bbox}{Bounding box for the region of interest.}

\item{fullDay}{Logical specifying whether to list files for an entire 
local time day.}

\item{daylightOnly}{Logical specifying whether to list only those files
containing daylight imagery somewhere over the region of interest specified
by \code{bbox}.}
}
\value{
Vector of filenames.
}
\description{
Retrieve a list of GOES AOD files available in the
\code{satelliteDataDir} for a specified date and hour.

This function is a wrapper for 
\preformatted{
files <- goesaodc_listFiles(
  satID = satID,
  datetime = datetime,
  timezone = timezone,
  bbox = bbox,
  fullDay = fullDay,
  daylightOnly = daylightOnly,
  useLocalDir = TRUE,
  baseUrl = baseUrl
) 
}

Note that all files for a particular hour will be returned even if the
incoming \code{startdate} or \code{jdate} is specified to the minute or
second.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

noon <- "2019-09-06 12:00"

# Default to noon UTC
goesaodc_listLocalFiles(
  satID = "G16", 
  datetime = datetime
)

# Noon on the west coast
goesaodc_listLocalFiles(
  satID = "G16", 
  datetime = datetime,
  timezone = "America/Los_Angeles"
)

# Full local time day (daylight anywhere in the US)
goesaodc_listLocalFiles(
  satID = "G16", 
  datetime = datetime,
  timezone = "America/Los_Angeles",
  fullDay = TRUE
)

CA_bbox <- c(-125, -114, 32, 42)

# Full local time day (daylight only in California)
goesaodc_listLocalFiles(
  satID = "G16", 
  datetime = datetime,
  timezone = "America/Los_Angeles",
  fullDay = TRUE,
  bbox = CA_bbox
)

}
}
\seealso{
\link{goesaodc_listFiles}

\link{goesaodc_listRemoteFiles}
}
