% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listScanFiles.R
\name{goesaodc_listScanFiles}
\alias{goesaodc_listScanFiles}
\title{List GOES AOD scan files}
\usage{
goesaodc_listScanFiles(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  isJulian = FALSE,
  useRemote = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/"
)
}
\description{
Lists GOES AOD scan files for a given datetime or time range. If
just \code{datetime} is given, then only the scan file closest to that time
will be listed. If \code{endtime} is specified as well, then all scans from
\code{datetime} up to (but not including) \code{endtime} will be listed.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:30",
  timezone = "UTC",
  useRemote = TRUE
)

goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "UTC",
  useRemote = TRUE
)
}
}
