% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_openFile.R
\name{goesaodc_openFile}
\alias{goesaodc_openFile}
\title{Open a GOES AOD netCDF file from the satelliteDataDir}
\usage{
goesaodc_openFile(fileName)
}
\arguments{
\item{fileName}{file name of GOES AOD file in the satelliteDataDir}
}
\value{
nc handle or NULL if no file was found
}
\description{
Open a GOES AOD netCDF file from the satelliteDataDir
}
