% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_openScanFile.R
\name{goesaodc_openScanFile}
\alias{goesaodc_openScanFile}
\title{Open a GOES AOD netCDF file}
\usage{
goesaodc_openScanFile(filename = NULL, dataDir = getSatelliteDataDir())
}
\arguments{
\item{filename}{Name of scan file.}

\item{dataDir}{Data directory containing \code{filename}. Defaults to 
\code{getSatelliteDataDir()}.}
}
\value{
NetCDF \code{ncdf4} handle.
}
\description{
Opens a NetCDF file and returns the \code{nc} handle. By 
default, the directory specified with \code{setSatelliteDataDir()} is 
searched for the file.
}
\details{
Errors generated by `nc_open()` are not handled and will pass through to the
console or calling function.
}
\examples{
\dontrun{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

scanFile <- goesaodc_downloadScanFiles(
  satID = "G16", 
  datetime = "2019-09-06 18:00", 
  timezone = "UTC"
)

goesaodc_openScanFile(scanFile)
}
}
\seealso{
\code{\link{getSatelliteDataDir}}
}
