% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotAODPoints.R
\name{goesaodc_plotAODPoints}
\alias{goesaodc_plotAODPoints}
\title{Create a point plot of averaged GOES AOD values}
\usage{
goesaodc_plotAODPoints(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = NULL,
  bbox = bbox_CONUS,
  dropNa = TRUE,
  breaks = NULL,
  cex = 0.5
)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Datetime in any Ymd H [MS] format or \code{POSIXct}.}

\item{endtime}{End time in any Ymd H [MS] format or \code{POSIXct} 
(exclusive).}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dropNa}{Logical flag whether to drop rows with NA values.}

\item{cex}{Plot symbol size.}
}
\description{
Averages the AOD values of each GOES sample location over 
multiple scans
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

# Kincade fire region
kincade_bbox <- c(-124, -120, 36, 39)

goesaodc_plotAODPoints(
  satID = "G16",
  datetime = "2019-10-27 10",
  timezone = "America/Los_Angeles",
  bbox = kincade_bbox,
  breaks = c(-Inf, 0, 1, 2, 3, 4, 5, Inf),
  cex = 0.6
)

maps::map(
  database = "state",
  regions = "california",
  xlim = c(-125, -120),
  ylim = c(36, 39),
  add  = TRUE
)
}
}
