% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotAveragePoints.R
\name{goesaodc_plotAveragePoints}
\alias{goesaodc_plotAveragePoints}
\title{Create a point plot of averaged GOES AOD values}
\usage{
goesaodc_plotAveragePoints(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = NULL,
  bbox = bbox_CONUS,
  dropNa = TRUE,
  breaks = NULL,
  cex = 0.5
)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Datetime in any Ymd H [MS] format or \code{POSIXct}.}

\item{endtime}{End time in any Ymd H [MS] format or \code{POSIXct} 
(exclusive).}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dropNa}{Logical flag whether to drop rows with NA values.}

\item{breaks}{Vector of color breaks.}

\item{cex}{Plot symbol size.}
}
\description{
Averages the AOD values of each GOES sample location over 
multiple scans
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

oregon_bbox <- c(-125, -116, 42, 47)

goesaodc_plotAveragePoints(
  satID = "G17",
  datetime = "2020-09-15 12",
  timezone = "America/Los_Angeles",
  bbox = oregon_bbox,
  cex = 0.3
)

maps::map(
  database = "state",
  regions = "oregon",
  xlim = oregon_bbox[1:2],
  ylim = oregon_bbox[3:4],
  add  = TRUE
)
}
}
