% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotScanRaster.R
\name{goesaodc_plotScanRaster}
\alias{goesaodc_plotScanRaster}
\title{Plot GOES AOD rasters}
\usage{
goesaodc_plotScanRaster(
  raster = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3,
  cellSize = NULL,
  fun = mean,
  rasterAlpha = NULL,
  paletteName = "YlOrRd",
  paletteBreaks = NULL,
  legendLimits = NULL,
  includeMap = FALSE,
  zoom = NULL,
  stateCodes = NULL,
  title = NULL
)
}
\arguments{
\item{raster}{A RasterBrick with AOD and DQF variables.}

\item{bbox}{Bounding box for the region of interest. Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}

\item{cellSize}{Size of Raster cells measured in degrees lon/lat.}

\item{fun}{Function used to summarize multiple point values within a single 
cell; Defaults to \code{mean}.}

\item{rasterAlpha}{Transparency of the raster. If not explicitly defined, it 
will default to 1.0 when \code{includeMap=FALSE} and 0.75 when
\code{includeMap=TRUE}.}

\item{paletteName}{The name of an RColorBrewer palette. Defaults to 'YlOrRd'.}

\item{paletteBreaks}{Vector of AOD values to use as palette breaks.}

\item{legendLimits}{Upper and lower AOD values for the fill legend. Setting 
this guarantees that the legend is displayed even if the scan has nothing but
NA AOD values. All values outside the range will be set to NA.}

\item{includeMap}{Logical flag to draw a topographic map image under the 
raster. Since the image is Mercator projected, the plot coordinate system 
will be Mercator projected to match. This significantly slows down the 
drawing of rasters. Defaults to FALSE.}

\item{zoom}{Zoom level of the topographic map, if it is included. Must be an
integer from 1 to 15.}

\item{stateCodes}{Codes of state outlines to draw.}

\item{title}{Title of the plot.}
}
\description{
Plot GOES AOD rasters
}
