% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotScanSPDF.R
\name{goesaodc_plotScanSPDF}
\alias{goesaodc_plotScanSPDF}
\title{Create a plot from a SpatialPointsDataFrame}
\usage{
goesaodc_plotScanSPDF(
  sp = NULL,
  bbox = bbox_CONUS,
  pointSize = 0.5,
  pointShape = 15,
  breaks = NULL,
  paletteName = "YlOrRd",
  title = NULL
)
}
\arguments{
\item{sp}{A SpatialPointsDataFrame.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{pointSize}{Size of plot points; Defaults to 0.5.}

\item{breaks}{Vector of AOD values to use as palette breaks.}

\item{paletteName}{The name of an RColorBrewer palette; Defaults to 'YlOrRd'.}

\item{title}{Title of the plot.}
}
\description{
Creates a plot of AOD points defined by a 
SpatialPointsDataFrame.
}
