% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotSpatialPoints.R
\name{goesaodc_plotSpatialPoints}
\alias{goesaodc_plotSpatialPoints}
\title{Create a quick plot of a GOES AOD SpatialPointsDataFrame}
\usage{
goesaodc_plotSpatialPoints(
  spatialPoints,
  var = "AOD",
  n = 1e+05,
  colBins = 5,
  breaks = NULL,
  paletteName = "YlOrRd",
  pch = 15,
  cex = 0.5,
  add = FALSE
)
}
\arguments{
\item{spatialPoints}{SpatialPointsDataFrame.}

\item{var}{GOES data variable ("AOD, "DQF" or "ID"); Defaults to "AOD".}

\item{n}{Sample size.}

\item{colBins}{Number of color bins.}

\item{breaks}{Vector of color breaks.}

\item{paletteName}{RColorBrewer palette name.}

\item{pch}{Plot character.}

\item{cex}{Plot symbol scale factor.}

\item{add}{Logical specifying whether to create a new plot or add to an 
existing one.}
}
\description{
Quickly subsamples and plots points in a GOES AOD 
spatialPointsDataFrame
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

netCDF <- goesaodc_downloadAOD(
  satID = "G17", 
  datetime = "2019-10-27 10", 
  timezone = "America/Los_Angeles" 
)[1]

# Kincade fire region
kincade_bbox <- c(-124, -120, 36, 39)

nc <- goesaodc_openFile(netCDF)
sp <- goesaodc_createSpatialPoints(nc, bbox = kincade_bbox, dqfLevel = 3)

goesaodc_plotSpatialPoints(sp, cex = 0.5)
maps::map(
  database = "state",
  regions = "california",
  xlim = c(-125, -119),
  ylim = c(35, 40),
  add  = TRUE
)
}
}
