% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_sampleScanRaster.R
\name{goesaodc_sampleScanRaster}
\alias{goesaodc_sampleScanRaster}
\title{Sample values from a scan raster}
\usage{
goesaodc_sampleScanRaster(
  raster = NULL,
  lon = NULL,
  lat = NULL,
  radius = 1,
  method = "simple",
  fun = mean,
  na.rm = FALSE
)
}
\arguments{
\item{raster}{\code{RasterLayer} with a variable.}

\item{lon}{Longitude position.}

\item{lat}{Latitude position.}

\item{radius}{Radius of a buffer around each point (in meters) from which to 
extract cell values. If the distance between the sampling point and the 
center of a cell is less than or equal to the buffer, the cell is included.}

\item{method}{Method of extracting cell values. If 'simple' values for the 
cell a point falls in are returned. If 'bilinear' the returned values are 
interpolated from the values of the four nearest raster cells. Defaults to 
'simple'.}

\item{fun}{Function used to summarize cell values if multiple cells fall
within the given radius. Defaults to `mean`.}

\item{na.rm}{Logical flag determining whether to remove \code{NA} values 
before summarizing them with `fun`. Defaults to \code{FALSE}.}
}
\value{
Numeric value extracted from the raster.
}
\description{
Sample values from a scan raster
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 46.5)

scanFile <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  timezone = "America/Los_Angeles"
)

# Create a raster for a scan file
scanRaster <- goesaodc_createScanRaster(
  filename = scanFile,
  bbox = bboxOregon,
  cellSize = 0.05
)

goesaodc_plotScanRaster(
  scanRaster,
  bbox = bboxOregon
)

goesaodc_sampleScanRaster(
  raster = scanRaster,
  lon = -123,
  lat = 46,
  radius = 10000
)
}
}
