% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSatelliteUtils.R
\docType{data}
\name{goesWestGrid}
\alias{goesWestGrid}
\title{GOES West Grid}
\format{A list with three elements}
\description{
This dataset is the latitude/longitude grid that GOES-17 ABI data is defined
on, computed from ABI scan angles provided in the netCDF and the projection
parameters found in \code{goesEastGrid$projection}. Because GOES-17 is a
geostationary satellite, the projection information should not change from
scan to scan. Using this dataset saves us from repeating the computationally
expensive task of converting scan angles to latitude/longitude.
}
\keyword{datasets}
