% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installGoesGrids.R
\name{installGoesGrids}
\alias{installGoesGrids}
\title{Create GOES grids in satelliteDataDir}
\usage{
installGoesGrids()
}
\arguments{
\item{nc}{ncdf4 object}
}
\description{
Creates data files with geolocation information for GOES-16 
(East) and GOES-17 (West) satellite products.

ROGER: Flesh out this documentation
}
