% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{isDaylight}
\alias{isDaylight}
\title{check if specified \code{datetime} is during the daylight hours of the
specified \code{region}}
\usage{
isDaylight(datetime = NULL, region = "CONUS")
}
\arguments{
\item{datetime}{desired datetime in any Y-m-d H [MS] format or \code{POSIXct}}

\item{region}{a named region. Currently only \code{"CONUS"} is supported}
}
\value{
logical
}
\description{
check if specified \code{datetime} is during the daylight hours of the
specified \code{region}
}
