% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goes.R
\name{isGoesProjection}
\alias{isGoesProjection}
\title{Check if `nc` projection is GOES projection}
\usage{
isGoesProjection(nc)
}
\arguments{
\item{nc}{ncdf4 object}
}
\value{
logical
}
\description{
Return TRUE if projection information in `nc` matches GOES projection
information.
}
