% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac_2nc4.R
\name{maiac_2nc4}
\alias{maiac_2nc4}
\title{Convert hdf file to .nc file}
\usage{
maiac_2nc4(file, converterPath = "./executables/h4toncff_nc4", cleanup = TRUE)
}
\arguments{
\item{file}{absolute path of file to be converted}

\item{converterPath}{location of converter function}

\item{cleanup}{logical specifying whether to remove original .hdf file}
}
\value{
Absolute path of converted NetCDF file.
}
\description{
Converts .hdf file to .nc file using h4toncff_nc4 http://www.hdfeos.org/software/h4cflib.php
}
\examples{
\dontrun{
hdfPath <- maiac_downloadNorthAmerica("h01v04", 20171009, 2150)
ncdfPath <- maiac_2nc4(hdfPath)
nc <- ncdf4::nc_open(ncdfPath)
nc
}
}
