% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac_cropRaw.R
\name{maiac_cropRaw}
\alias{maiac_cropRaw}
\title{Crop raw maiac tibble}
\usage{
maiac_cropRaw(maiac_rawDataframe, xlim = NULL, ylim = NULL)
}
\arguments{
\item{maiac_rawDataframe}{maiac dataframe generated from maiac_loadRawDataframe()}

\item{xlim}{low and high longitude values (xlo, xhi)}

\item{ylim}{low and high latitude values (ylo, yhi)}
}
\value{
cropped raw maiac tibble
}
\description{
crop a raw maiac dataframe
}
\examples{
\dontrun{
maiac <- maiac_loadRawDataframe("h01v04", 20171009, 2150)
maiac <- maiac_cropRaw(maiac, xlim = c(-125.52, -117.03), ylim = c(36.52, 41.78))
}
}
