% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac_downloadNorthAmerica.R
\name{maiac_downloadNorthAmerica}
\alias{maiac_downloadNorthAmerica}
\title{Download North America MAIAC swath data}
\usage{
maiac_downloadNorthAmerica(tileNumber = NULL, date = NULL,
  time = NULL, julianDate = NULL, product = "MAIACAAOT",
  baseUrl = "https://portal.nccs.nasa.gov/datashare/maiac/DataRelease/NorthAmerica_2000-2016/")
}
\arguments{
\item{tileNumber}{number code for tile (eg h01v04).}

\item{date}{desired date (integer, character representing YYYYMMDD[HH] or datetime object)}

\item{time}{UTC hour and minute for data (HHMM). Optional when only one file for the specified date.}

\item{julianDate}{desired date on Julian calendar (YYYYDDD). Ignored if 'date' is specified.}

\item{product}{product code (MAIACAAOT | MAIACABRF | MAIACRTLS | MAIACTAOT | MAIACTBRF)}

\item{baseUrl}{base URL for data queries}
}
\value{
0 if successful.
}
\description{
Download .hdf file of a MAIAC tile for a particular day.
}
\examples{
\dontrun{
maiac_downloadNorthAmerica( "h01v04", 20171009, 2150 )
}
}
\references{
\href{http://usfs.airsis.com}{Interagency Real Time Smoke Monitoring}
}
