% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac_getMetadata.R
\name{maiac_getMetadata}
\alias{maiac_getMetadata}
\title{Get metadata list from metadata string in .nc file}
\usage{
maiac_getMetadata(filePath)
}
\arguments{
\item{filePath}{path to .nc file}
}
\value{
metadata list
}
\description{
Get metadata list from metadata string in .nc file
}
\examples{
\dontrun{
hdfPath <- maiac_downloadNorthAmerica("h01v04", 20171009, 2150)
ncdfPath <- maiac_2nc4(hdfPath)
maiac_getMetadata(ncdfPath)
}
}
