% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac_loadRaster.R
\name{maiac_loadRaster}
\alias{maiac_loadRaster}
\title{Download North America MAIAC swath data}
\usage{
maiac_loadRaster(tileNumber = NULL, date = NULL, time = NULL,
  julianDate = NULL, product = "MAIACAAOT",
  baseUrl = "https://portal.nccs.nasa.gov/datashare/maiac/DataRelease/NorthAmerica_2000-2016/",
  filePath = NULL, params = "Optical_Depth_055",
  converterPath = "./executables/h4toncff_nc4")
}
\arguments{
\item{tileNumber}{number code for tile (eg h01v04).}

\item{date}{desired date (integer, character representing YYYYMMDD[HH] or datetime object)}

\item{time}{UTC hour and minute for data (HHMM). Optional when only one file for the specified date.}

\item{julianDate}{desired date on Julian calendar (YYYYDDD). Ignored if 'date' is specified.}

\item{product}{product code (MAIACAAOT | MAIACABRF | MAIACRTLS | MAIACTAOT | MAIACTBRF)}

\item{baseUrl}{base URL for data queries}

\item{filePath}{optional path of .nc file. If not null, all other arguments are ignored.}

\item{params}{vector of variable names to use for the values of the grid cells. Grid size must match. 
"all_1km" or "all_5km" may also be used to include layers for all parameters on the 1km grid or 5km grid.}

\item{converterPath}{Path to executable for converting HDF4 to netCDF4}
}
\value{
raster::Raster* object
}
\description{
Load North America MAIAC swath data as a raster
}
\examples{
\dontrun{
napaSmoke <- maiac_loadRaster( "h01v04", 20171009, 2150 )
plot(napaSmoke)
}
}
