% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac_loadRawDataframe.R
\name{maiac_loadRawDataframe}
\alias{maiac_loadRawDataframe}
\title{Download or load MAIAC data and return data as a dataframe}
\usage{
maiac_loadRawDataframe(tileNumber = NULL, date = NULL, time = NULL,
  julianDate = NULL, product = "MAIACAAOT", baseUrl = NULL)
}
\arguments{
\item{tileNumber}{number code for tile (eg h01v04).}

\item{date}{desired date (integer, character representing YYYYMMDD[HH] or datetime object)}

\item{time}{UTC hour and minute for data (HHMM). Optional when only one file for the specified date.}

\item{julianDate}{desired date on Julian calendar (YYYYDDD). Ignored if 'date' is specified.}

\item{product}{product code (MAIACAAOT | MAIACABRF | MAIACRTLS | MAIACTAOT | MAIACTBRF)}

\item{baseUrl}{base URL for data queries}
}
\value{
Raw tibble of MAIAC data
}
\description{
load a raw dataframe
}
\examples{
\dontrun{
maiac <- maiac_loadRawDataframe("h01v04", 20171009, 2150)
}
}
