% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiac_pm25ToAod.R
\name{maiac_pm25ToAOD}
\alias{maiac_pm25ToAOD}
\title{Convert column-integrated PM2.5 values to AOD}
\usage{
maiac_pm25ToAOD(pm25)
}
\arguments{
\item{pm25}{vector of column-integrated pm25 values}
}
\value{
vector of AOD values.
}
\description{
Converts a vector of PM2.5 values to AOD values
}
\keyword{utils}
