% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSatelliteUtils.R
\docType{package}
\name{MazamaSatelliteUtils}
\alias{MazamaSatelliteUtils}
\title{Utilities for working with model and satellite data as rasters.}
\source{
\url{https://www.ncdc.noaa.gov/sites/default/files/attachments/GOES-17_ABI_L2_AOD_Provisional_ReadMe.pdf}

The package contains two sample NetCDF files with GOES-16 and GOES-17 AOD 
data from for 2019-09-06 18:26 UTC -- 1pm Pacific Daily Time.

Access these files with:

\preformatted{
G16_filepath <- system.file(
  "extdata", 
  "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc", 
  package = "MazamaSatelliteUtils")

G17_filepath <- system.file(
  "extdata", 
  "OR_ABI-L2-AODC-M6_G17_s20192491826196_e20192491828569_c20192491830494.nc", 
  package = "MazamaSatelliteUtils")
}
}
\description{
A suite of utility functions providing functionality commonly 
needed for download of satellite products and conversion to Raster objects.
}
\section{GOES AOD}{

ABI L2+ AOD NetCDF files from NOAA contain the aerosol optical 
depth at 550 nm over land and over water, associated quality flags, mean, 
maximum, minimum and standard deviation of 550-nm AOD for the entire domain. 
The AOD is a measure of the columnar extinction (scattering +absorption) of 
radiation by aerosols. It is proportional to the amount (number or mass 
concentration) of aerosols in an atmospheric column. 
\describe{
  \item{dqf}{data quality}
  \item{aod}{measure optical depth}
  ...
}
}

