% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_createDaytimeRasterStack.R
\name{ra_createDaytimeRasterStack}
\alias{ra_createDaytimeRasterStack}
\title{Create a daytime RasterStack for a specified date}
\usage{
ra_createDaytimeRasterStack(satID = NULL, datetime = NULL,
  longitude = NULL, latitude = NULL, var = "AOD", res = 0.1,
  bbox = NULL, dqfLevel = NULL, timezone = "UTC")
}
\description{
ra_createDaytimeRasterStack.R
}
\examples{
\dontrun{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)
setSatelliteDataDir("~/Data/Satellite")
MazamaCoreUtils::initializeLogging(logDir = "D:Mazama/Data/Logs")

# Define the region of interest (Milepost 97 Fire in Oregon)
bbox_oregon <- c(-124.56624, -116.46350, 41.99179, 46.29203)
longitude <- -123.245
latitude <-   42.861

datetime <- MazamaCoreUtils::parseDatetime(datetime = "2019-09-06", 
timezone = "America/Los_Angeles")

dayStack <- ra_createDaytimeRasterStack(
  satID = "G16",
  datetime = datetime,
  longitude = longitude,
  latitude = latitude,
  bbox = bbox_oregon
)

tb <- raster_createLocationTimeseries(dayStack,
                                      longitude = longitude,
                                      latitude = latitude,
                                      bbox = bbox_oregon)

plot(x = tb$datetime, y = tb$aod,
     pch = 15, cex = 0.8, col = rgb(red = 0, green = 0, blue = 0, alpha = 0.8),
     main = datetime, xlab = "Time (PDT)", ylab = "AOD")
}
}
