% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaSatelliteUtils.R
\name{setSatelliteDataDir}
\alias{setSatelliteDataDir}
\title{Set Satellite Data Directory}
\usage{
setSatelliteDataDir(dataDir)
}
\arguments{
\item{dataDir}{directory where Satellite datasets are stored}
}
\value{
Silently returns previous value of data directory.
}
\description{
Sets the data directory where Satellite data is located.
If the directory does not exist, it will be created.
}
\seealso{
\link{SatelliteDataDir}

\link{getSatelliteDataDir}
}
\keyword{environment}
