% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bboxToVector}
\alias{bboxToVector}
\title{Converts and validates bbox objects or strings into standard format}
\usage{
bboxToVector(bbox = NULL)
}
\arguments{
\item{bbox}{bounding box for region of interest.}
}
\value{
a vector of floats in c(lonLo, lonHi, latLo, latHi) order.
}
\description{
Accepts a variety of bbox formats, validates that they are
numerically sane and returns a vector of floats in c(lonLo, lonLo, latLo, latHi) order.
Input can be a vector of floats in c(lonLo, lonHi, latLo, latHi) order or 
the return value from \code{sp::bbox()} or \code{raster::extent()}.
}
\examples{
\donttest{
# USE SAMPLE CODE FROM isDaylight.R
}

}
