% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bboxToVector}
\alias{bboxToVector}
\title{Converts and validates bbox objects or strings into standard format}
\usage{
bboxToVector(bbox = NULL)
}
\arguments{
\item{bbox}{bounding box for region of interest.}
}
\value{
a vector of floats in c(lonLo, lonHi, latLo, latHi) order.
}
\description{
Accepts a variety of bbox formats, validates that they are
numerically sane and returns a vector of floats in c(lonLo, lonLo, latLo, latHi) order.
Input can be a vector of floats in c(lonLo, lonHi, latLo, latHi) order or 
the return value from \code{sp::bbox()} or \code{raster::extent()}.
}
\examples{
\donttest{
library(MazamaSpatialUtils)
library(MazamaSatelliteUtils)

setSpatialDataDir("~/Data/Spatial")
loadSpatialData("USCensusStates")
setSatelliteDataDir("~/Data/Satellite")

oregon <- subset(USCensusStates, stateCode == "OR")
bbox <- sp::bbox(oregon)

ncFile <- system.file(
  "extdata",
  "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc",
  package = "MazamaSatelliteUtils"
)
nc <- goesaodc_openFile(ncFile)

raster <- goesaodc_createRaster(nc, res = 0.1, dqfLevel = 2)
extent <- raster::extent(raster)

# Convert bbox
bbox
bboxToVector(bbox)

# Convert extent
extent
bboxToVector(extent)

# Convert low/hi mixup
bboxToVector(c(-116, -124, 42, 46))
}

}
