% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGoesGrid.R
\name{createGoesGrid}
\alias{createGoesGrid}
\title{Create a GOES grid in satelliteDataDir}
\usage{
createGoesGrid(nc = NULL, grid_filepath = NULL)
}
\arguments{
\item{nc}{ncdf4 handle.}

\item{grid_filepath}{filepath location for either goesEastGrid.rda or 
goesWestGrid.rda.}
}
\value{
Invisibly returns the file path of the created GOES grid file.
}
\description{
Creates data files with geolocation information for GOES-16 
(East) or GOES-17 (West) satellite products. Takes an open .nc filehandle 
and reads projection and coordinate grid information from it to create
a GOES East or West grid in the directory previously set with 
\code{setSatelliteDataDir()}.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

scanFilename <- goesaodc_listScanFiles(
  satID = "G16", 
  datetime = "2019-9-6 18:00",
  timezone = "UTC",
)
   
nc <- goesaodc_openScanFile(scanFilename)
G16_filepath <- file.path(getSatelliteDataDir(), "goesEastGrid.rda")

createGoesGrid(nc, G16_filepath)
} 
}
