% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{getAttributes}
\alias{getAttributes}
\title{Returns attributes from a netCDF file}
\usage{
getAttributes(nc, varid = 0, verbose = TRUE)
}
\arguments{
\item{nc}{ncdf4 handle.}

\item{varid}{The variable whose attribute is to be read. Can be a character 
string with the variable's name or an object of class ncvar4. As a special 
case, if \code{varid == 0}, then a global (file) attribute will be read 
rather than a particular variable's attribute.}

\item{verbose}{Logical flag to print attribute structure. Defaults to TRUE.}
}
\value{
Invisibly returns a named list of netCDF attributes.
}
\description{
Returns attributes from a netCDF file.
}
