% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDaylightHours.R
\name{getDaylightHours}
\alias{getDaylightHours}
\title{Calculates the sunrise and sunset hours for a specific date and 
location}
\usage{
getDaylightHours(datetime = NULL, longitude = NULL, latitude = NULL,
  bbox = NULL, timezone = NULL, julian = FALSE)
}
\arguments{
\item{datetime}{Specific date for which daylight hours are requested.}

\item{longitude}{Longitude of the location of interest in decimal degrees E.}

\item{latitude}{Latitude of the location of interest in decimal degrees N.}

\item{bbox}{Bounding box for the region of interest, Default: CONUS.}

\item{timezone}{Timezone in which to interpret the \code{datetime}.}

\item{julian}{Logical value determining whether datetime should be 
interpreted as a Julian date with day of year as a decimal number.}
}
\value{
List containing \code{POSIXct} \code{sunrise} and \code{sunset} times 
for the specified date.
}
\description{
Returns the sunrise and sunset hours for specified dates within 
a bounding box based on sun ephemerides calculations. Sunrise is calcualted
using the midpoint of the east edge of the bounding box while sunset is
calculated using the midpoint of the west edge.

The bounding box may be specified via any of the following methods:

\itemize{
\item{}{Olson timezone, either as part of a POSIXt datetime, or specified 
via the \code{timezone} parameter}
\item{}{Geographic Bounding Box, either as a vector of floats in 
\code{c(lonLo, lonHi, latLo,latHi)} order, or the return value from 
\code{sp::bbox()}, or \code{raster::extent()} }
\item{}{Longitude and Latitude specified in decimal degrees}
}

If more than one of the above are specified, the order of preference for 
determining the bounding box is:

\enumerate{
\item{\code{longitude, latitude} passed in -- location +/- 0.1 degrees}
\item{\code{bbox} passed in -- use unchanged}
\item{\code{timezone} passed in -- use bbox of the timezone}
\item{\code{datetime} is \code{POSIXct} -- use bbox of the timezone attribute}
}
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

# EX 1. Get Sunrise and Sunset by Lat/Lon and Datetime
daylight <- getDaylightHours(
  datetime = "2019-09-06", 
  longitude = -123.245, 
  latitude = 42.861
)

# EX 2. Get Sunrise and Sunset by BBOX and datetime
daylight <- getDaylightHours(
  datetime = "2019-09-06",
  bbox = c(-124.566, -116.463, 41.991, 46.292)
) # Oregon

# EX 3. Get Sunrise and Sunset by Datetime and Timezone
daylight <- getDaylightHours(
  datetime = "2019-09-06",
  timezone = "America/Los_Angeles")
}

}
