% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{getValue}
\alias{getValue}
\title{Get raster value at a point}
\usage{
getValue(raster, x, y)
}
\arguments{
\item{raster}{\code{RasterLayer} object}

\item{x}{longitude value (deg E)}

\item{y}{latitude value (deg N)}
}
\value{
Value of raster cell covering the point.
}
\description{
Get raster value at a point.
}
\keyword{utils}
