% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{getValue}
\alias{getValue}
\title{Get raster value at a point}
\usage{
getValue(raster, x, y)
}
\arguments{
\item{raster}{A \code{RasterLayer} object.}

\item{x}{Longitude value (deg E).}

\item{y}{Latitude value (deg N).}
}
\value{
Value of raster cell covering the point.
}
\description{
Gets the raster value at a given point.
}
\keyword{utils}
