% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goes_createTibble.R
\name{goes_createTibble}
\alias{goes_createTibble}
\title{Create a tibble from ncdf4 handle}
\usage{
goes_createTibble(nc, vars = c("AOD", "DQF"))
}
\arguments{
\item{nc}{ncdf4 handle}

\item{vars}{list of variables}
}
\value{
tibble
}
\description{
Create a tibble with columns: AOD, DQF, lon, and lat
}
