% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_areaPlot.R
\name{goesaodc_areaPlot}
\alias{goesaodc_areaPlot}
\title{Create plots of AOD data available within a region}
\usage{
goesaodc_areaPlot(ncList = NULL, bbox = bbox_CONUS, dqfLevel = 2,
  exponentiate = TRUE, col_state = "black", col_county = "white",
  lwd_state = 1.5, lwd_county = 1)
}
\arguments{
\item{ncList}{ncdf4 handle or a list of handles.}

\item{bbox}{Geographic extent of area of interest; Defaults to CONUS.}

\item{dqfLevel}{Sets the DQF level to filter to data to.}

\item{exponentiate}{Logical, sets whether to scale the data values by 10^AOD}

\item{col_state}{Color of state borders. Use "transparent" for no lines.}

\item{col_county}{Color of county borders. Use "transparent" for no lines.}

\item{lwd_state}{Line weight of state borders.}

\item{lwd_county}{Line weight of county borders.}
}
\description{
The goal of this plot is to get a quick look at available data 
within a region. A user might look at a single timestep or might pass in a 
list of nc handles to see if native grid averaging over several time steps 
significantly reduces the number missing data grid cells.
}
\examples{
\dontrun{
library(sp)
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("USCensusStates")

library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

# Date and region of interest
files <- goesaodc_downloadAOD(satID = "G17", 
  datetime = 2019102714, 
  timezone = "America/Los_Angeles")[1]
  
#' Kincade fire
bbox <- c(-124, -120, 36, 39)

# Build a list of open nc handles to process
ncList <- list()
for ( file in files ) {
  label <- 
    file \%>\%
    goesaodc_convertFilenameToDatetime() \%>\%
    MazamaCoreUtils::timeStamp(unit = "sec", timezone = "UTC")
  ncList[[label]] <- goesaodc_openFile(basename(file))
}

goesaodc_areaPlot(ncList, kincade_bbox)
}
}
