% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_convertFilenameToDatetime}
\alias{goesaodc_convertFilenameToDatetime}
\title{Get the scan start time from a GOES AODC netCDF file name}
\usage{
goesaodc_convertFilenameToDatetime(file)
}
\arguments{
\item{file}{GOES AOD netCDF file name}
}
\value{
The scan start time.
}
\description{
Get the scan start time from a GOES AODC netCDF file name
}
