% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createDaytimeRasterStack.R
\name{goesaodc_createDaytimeRasterStack}
\alias{goesaodc_createDaytimeRasterStack}
\title{Create a daytime RasterStack for a specified date}
\usage{
goesaodc_createDaytimeRasterStack(satID = NULL, datetime = NULL,
  longitude = NULL, latitude = NULL, var = "AOD", res = 0.1,
  bbox = NULL, dqfLevel = NULL, timezone = "UTC")
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Desired datetime in any Ymd H [MS] format or \code{POSIXct}.}

\item{longitude}{Target longitude used to calcualte daylight hours.}

\item{latitude}{Target latitude used to calculate daylight hours.}

\item{var}{variable ("AOD, "DQF" or "ID")}

\item{res}{resolution of raster in degrees}

\item{bbox}{Bounding box for the region of interest.}

\item{dqfLevel}{Data quality flag level.}

\item{timezone}{timezone in which to interpret the \code{datetime}.}
}
\value{
RasterStack
}
\description{
Create a \code{RasterStack} from GOES AOD data files for the
date specified by \code{datetime}. Each \code{RasterLayer} contains
data from one Advanced Baseline Imager (ABI) scan during the specified time
period.

If data for the specified time period is not found in the directory specified
by \code{setSatelliteDataDir()}, it will be downloaded in order to create the
\code{RasterStack}.

The Z axis of the \code{RasterStack} is a character vector where each element
is the time stamp of the scan and has the format YYYYMMDDHHMMSS. This can be
accessed using the \code{raster::getZ()} function. Names of the
\code{RasterStack} are also time stamps of the scan, of the format XHH.MM.SS.

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or 
\code{raster::extent()}.
}
\examples{
\dontrun{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")
MazamaCoreUtils::initializeLogging(logDir = "~/Data/Logs")

# Define the region of interest (Milepost 97 Fire in Oregon)
bbox_oregon <- c(-124.56624, -116.46350, 41.99179, 46.29203)
longitude <- -123.245
latitude <-   42.861

datetime <- MazamaCoreUtils::parseDatetime(datetime = "2019-09-06", 
timezone = "America/Los_Angeles")

dayStack <- goesaodc_createDaytimeRasterStack(
  satID = "G16",
  datetime = datetime,
  longitude = longitude,
  latitude = latitude,
  bbox = bbox_oregon
)

tb <- raster_createLocationTimeseries(dayStack,
                                      longitude = longitude,
                                      latitude = latitude,
                                      bbox = bbox_oregon)

plot(x = tb$datetime, y = tb$aod,
     pch = 15, cex = 0.8, col = rgb(red = 0, green = 0, blue = 0, alpha = 0.8),
     main = datetime, xlab = "Time (PDT)", ylab = "AOD")
}
}
