% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createDaytimeRasterStack.R
\name{goesaodc_createDaytimeRasterStack}
\alias{goesaodc_createDaytimeRasterStack}
\title{Create a daytime RasterStack for a specified date}
\usage{
goesaodc_createDaytimeRasterStack(startdate = NULL, longitude = NULL,
  latitude = NULL, var = "AOD", res = 0.1, fun = mean,
  bbox = NULL, lonLo = NULL, lonHi = NULL, latLo = NULL,
  latHi = NULL, dqfLevel = NULL)
}
\arguments{
\item{startdate}{startdate in any Y-m-d format or \code{POSIXct}}

\item{var}{variable ("AOD, "DQF" or "ID")}

\item{res}{resolution of raster in degrees}

\item{fun}{function to use when rasterizing. Not currently supported, defaults
to mean.}

\item{bbox}{bounding box generated by \code{sp::bbox}. Overrides lonLo, lonHi
latLo, latHi if provided}

\item{lonLo}{lower longitude extent}

\item{lonHi}{upper longitude extent}

\item{latLo}{lower latitude extent}

\item{latHi}{upper latitude extent}

\item{dqfLevel}{data quality flag level, numeric (0, 1, 2, or 3)}

\item{timezone}{the timezone to determine the sunrise and sunset times}
}
\value{
RasterStack
}
\description{
Create a \code{RasterStack} from GOES AOD data files for the 
date specified by \code{startdate}. Each \code{RasterLayer} contains
data from one Advanced Baseline Imager (ABI) scan during the specified time 
period.

If data for the specified time period is not found in the directory specified 
by \code{setSatelliteDataDir()}, it will be downloaded in order to create the
\code{RasterStack}.

The Z axis of the \code{RasterStack} is a character vector where each element
is the time stamp of the scan and has the format YYYYMMDDHHMMSS. This can be 
accessed using the \code{raster::getZ()} function. Names of the 
\code{RasterStack} are also time stamps of the scan, of the format XHH.MM.SS.

Data quality level `dqfLevel` can take a value of:

0: High quality retrieval flag
1: Medium quality retrieval flag
2: Low quality retrieval flag
3: No retrieval quality flag
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("USCensusStates")

# Define the region of interest (Milepost 97 Fire in Oregon)
oregon <- subset(USCensusStates, stateCode == "OR")
bbox_oregon <- sp::bbox(oregon)
lon <- -123.245
lat <-   42.861

dateLocal <- lubridate::ymd("2019-08-01", tz = "America/Los_Angeles")

dayStack <- goesaodc_createDaytimeRasterStack(dateLocal, 
                                              longitude = -123.32,
                                              latitude = 42.88,
                                              bbox = bbox_oregon)
tb <- raster_createLocationTimeseries(dayStack, 
                                      longitude = lon, latitude = lat, 
                                      bbox = bbox_oregon)

plot(x = tb$datetime, y = tb$aod, 
     pch = 15, cex = 0.8, col = rgb(red = 0, green = 0, blue = 0, alpha = 0.8),
     main = dateLocal, xlab = "Time (PDT)", ylab = "AOD")
}
}
