% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createDaytimeRasterStack.R
\name{goesaodc_createDaytimeRasterStack}
\alias{goesaodc_createDaytimeRasterStack}
\title{Create a daytime RasterStack for a specified date}
\usage{
goesaodc_createDaytimeRasterStack(satID = NULL, datetime = NULL,
  var = "AOD", res = 0.1, bbox = NULL, dqfLevel = NULL,
  timezone = NULL, verbose = FALSE)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Desired datetime in any Ymd H [MS] format or \code{POSIXct}.}

\item{var}{variable ("AOD, "DQF" or "ID")}

\item{res}{resolution of raster in degrees}

\item{bbox}{Bounding box for the region of interest.}

\item{dqfLevel}{Data quality flag level.}

\item{timezone}{timezone in which to interpret the \code{datetime}.}

\item{verbose}{show progress of raster stacking.}
}
\value{
RasterStack
}
\description{
Create a \code{RasterStack} from GOES AOD data files for the
date specified by \code{datetime}. Each \code{RasterLayer} contains
data from one Advanced Baseline Imager (ABI) scan during the specified time
period.

If data for the specified time period is not found in the directory specified
by \code{setSatelliteDataDir()}, it will be downloaded in order to create the
\code{RasterStack}.

The Z axis of the \code{RasterStack} is a character vector where each element
is the time stamp of the scan and has the format YYYYMMDDHHMMSS. This can be
accessed using the \code{raster::getZ()} function. Names of the
\code{RasterStack} are also time stamps of the scan, of the format XHH.MM.SS.

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or 
\code{raster::extent()}.
}
\examples{
\dontrun{

library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("USCensusStates")

calif <- subset(USCensusStates, stateCode == "CA")

bbox_ca   <- sp::bbox(calif)
satID     <- "G16"
datetime  <- "2019-10-23"
timezone  <- "America/Los_Angeles"
dqfLevel  <- 2
latitude  <- 38.448611 
longitude <- -122.704722

dayStack <- goesaodc_createDaytimeRasterStack(
  satID = satID,
  datetime = datetime,
  timezone = timezone,
  bbox = bbox_ca,
  dqfLevel = dqfLevel
)

tb <- raster_createLocationTimeseries(dayStack,
                                      longitude = longitude,
                                      latitude = latitude,
                                      bbox = bbox_ca)

plot(x = tb$datetime, y = tb$aod,
     pch = 15, cex = 0.8, col = rgb(red = 0, green = 0, blue = 0, alpha = 0.8),
     main = "Santa Rosa, 2019-10-23", xlab = "Time (UTC)", ylab = "AOD")
}
}
