% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createEmptyScanPoints.R
\name{goesaodc_createEmptyScanPoints}
\alias{goesaodc_createEmptyScanPoints}
\title{Create an empty GOES scan SPDF}
\usage{
goesaodc_createEmptyScanPoints(
  satID = NULL,
  filename = NULL,
  bbox = bbox_CONUS
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest. Defaults to CONUS.}
}
\description{
Creates a SpatialPointsDataFrame of AOD readings from one or 
more GOES scans. Scans are specified either by filename or satellite + time 
information.
}
