% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createEmptyScanSpdf.R
\name{goesaodc_createEmptyScanSpdf}
\alias{goesaodc_createEmptyScanSpdf}
\title{Create an empty SPDF for a GOES satellite}
\usage{
goesaodc_createEmptyScanSpdf(satID = NULL, filename = NULL, bbox = bbox_CONUS)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest. Defaults to CONUS.}
}
\value{
A \code{SpatialPointsDataFrame} filled with \code{NA} AOD values.
}
\description{
Creates a \code{SpatialPointsDataFrame} of \code{NA} AOD 
readings for a GOES satellite. This is useful when trying to create an SPDF 
for a faulty scan file.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

MazamaSatelliteUtils:::goesaodc_createEmptyScanSpdf(satID = "G16")
MazamaSatelliteUtils:::goesaodc_createEmptyScanSpdf(satID = "G17")
MazamaSatelliteUtils:::goesaodc_createEmptyScanSpdf(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202530031174_e20202530033547_c20202530035523.nc"
)
}
}
