% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tate-dailyAverage.R
\name{goesaodc_createHourlyAverageRaster}
\alias{goesaodc_createHourlyAverageRaster}
\title{Create a RasterBrick Average for a specified hour}
\usage{
goesaodc_createHourlyAverageRaster(startdate = NULL, state = NULL,
  res = 0.1, dqfLevel = NULL, na.rm = TRUE)
}
\arguments{
\item{startdate}{startdate, specified to the hour, in any Y-m-d H format or
\code{POSIXct}}

\item{res}{resolution of raster in degrees}

\item{dqfLevel}{data quality flag level, numeric (0, 1, 2, or 3)}

\item{bbox}{bounding box generated by \code{sp::bbox}.}
}
\value{
RasterBrick
}
\description{
Create 

Data quality level `dqfLevel` can take a value of:

0: High quality retrieval flag
1: Medium quality retrieval flag
2: Low quality retrieval flag
3: No retrieval quality flag
}
\examples{
\donttest{
}
}
