% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createHourlyRasterStack.R
\name{goesaodc_createHourlyRasterStack}
\alias{goesaodc_createHourlyRasterStack}
\title{Create a RasterStack of all files from the specified hour}
\usage{
goesaodc_createHourlyRasterStack(datetime = NULL, var = "AOD",
  res = 0.1, fun = mean, bbox = NULL, lonLo = NULL, lonHi = NULL,
  latLo = NULL, latHi = NULL, dqfLevel = NULL)
}
\arguments{
\item{datetime}{datetime, specified to the hour, in any Y-m-d H format or
\code{POSIXct}}

\item{var}{variable ("AOD, "DQF" or "ID")}

\item{res}{resolution of raster in degrees}

\item{fun}{function to use when rasterizing. Not currently supported, defaults
to mean.}

\item{bbox}{bounding box generated by \code{sp::bbox}. Overrides lonLo, lonHi
latLo, latHi if provided}

\item{lonLo}{lower longitude extent}

\item{lonHi}{upper longitude extent}

\item{latLo}{lower latitude extent}

\item{latHi}{upper latitude extent}

\item{dqfLevel}{data quality flag level, numeric (0, 1, 2, or 3)}
}
\value{
RasterStack
}
\description{
Create a RasterStack of GOES AOD data of all files for a
specified hour. 

Data quality level can take a value of:

0: High quality retrieval flag
1: Medium quality retrieval flag
2: Low quality retrieval flag
3: No retrieval quality flag
}
