% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createHourlyRasterStack.R
\name{goesaodc_createHourlyRasterStack}
\alias{goesaodc_createHourlyRasterStack}
\title{Create a RasterStack for a specified hour}
\usage{
goesaodc_createHourlyRasterStack(satID = NULL, datetime = NULL,
  var = "AOD", res = 0.1, fun = mean, bbox = NULL,
  dqfLevel = NULL, timezone = "UTC")
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{desired datetime in any Ymd H [MS] format or \code{POSIXct}}

\item{var}{variable ("AOD, "DQF" or "ID")}

\item{res}{resolution of raster in degrees}

\item{fun}{function to use when rasterizing. Not currently supported, 
defaults to mean.}

\item{bbox}{Bounding box for the region of interest.}

\item{dqfLevel}{Data quality flag level.}

\item{timezone}{timezone in which to interpret the \code{datetime}.}
}
\value{
RasterStack
}
\description{
Create a \code{RasterStack} from GOES AOD data files for the 
date and hour specified by \code{datetime}. Each \code{RasterLayer} contains
data from one Advanced Baseline Imager (ABI) scan during the specified time 
period.

If data for the specified time period is not found in the directory specified 
by \code{setSatelliteDataDir()}, it will be downloaded in order to create the
\code{RasterStack}.

The Z axis of the \code{RasterStack} is a character vector where each element
is the time stamp of the scan and has the format YYYYMMDDHHMMSS. This can be 
accessed using the \code{raster::getZ()} function. Names of the 
\code{RasterStack} are also time stamps of the scan, of the format XHH.MM.SS.

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or 
\code{raster::extent()}.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")
rasterStack <- goesaodc_createHourlyRasterStack(
  satID = "G16", 
  datetime = "2019-05-16 16:00", 
  res = 0.2
)
rasterVis::levelplot(rasterStack)
}
}
