% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createNativeGrid.R
\name{goesaodc_createNativeGrid}
\alias{goesaodc_createNativeGrid}
\title{Create a nativeGrid object}
\usage{
goesaodc_createNativeGrid(nc = NULL, bbox = bbox_CONUS)
}
\arguments{
\item{nc}{ncdf4 handle or a list of handles.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}
}
\value{
List with lon, lat, AOD and DQF matrices
}
\description{
Creates a \emph{nativeGrid} object with a list of matrices for: 
longitude, latitude, AOD values, and DQF values. The arrays are defined in
native i, j coordinates and are thus curvilinear as opposed to rectilinear 
geospatial coordinates.

The \code{nc} parameter can be either a single netcdf handle or a list of 
handles. If a list of handles is provided, grid cell values for AOD and DQF 
will be averaged across all nc handles. This "native grid" averaging 
provides a simple way to convert 5-minute data into an hourly average.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadScanFiles(
  satID = "G17", 
  datetime = "2019-10-27 14:00",
  endtime = "2019-10-27 15:00",
  timezone = "America/Los_Angeles"
)

scanFiles <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2019-10-27 14:00",
  endtime = "2019-10-27 15:00",
  timezone = "America/Los_Angeles"
)

ncList <- goesaodc_openScanFile(scanFiles)

# Kincade fire region
kincade_bbox <- c(-124, -120, 36, 39)

layout(matrix(seq(2)))

nativeGrid <- goesaodc_createNativeGrid(ncList[1], kincade_bbox)
image(nativeGrid$AOD[,ncol(nativeGrid$AOD):1])
title("Single timestep")

nativeGrid <- goesaodc_createNativeGrid(ncList, kincade_bbox)
image(nativeGrid$AOD[,ncol(nativeGrid$AOD):1])
title("Average of 12 timesteps")

layout(1)
}
}
