% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createRaster.R
\name{goesaodc_createRaster}
\alias{goesaodc_createRaster}
\title{Create a RasterBrick of GOES data}
\usage{
goesaodc_createRaster(nc, res = 0.1, fun = mean, xmn = NULL,
  xmx = NULL, ymn = NULL, ymx = NULL, dqfLevel = NULL)
}
\arguments{
\item{nc}{ncdf4 handle}

\item{res}{resolution of raster in degrees}

\item{fun}{function to use when rasterizing}

\item{xmn}{lower longitude extent}

\item{xmx}{upper longitude extent}

\item{ymn}{lower latitude extent}

\item{ymx}{upper latitude extent}

\item{dqfLevel}{data quality flag level}
}
\value{
RasterBrick
}
\description{
Create a RasterBrick of GOES AOD data including data points
with the specified resolution and function, and within the specified extent 
and data quality flag level. Data quality level can take a value of:

0: High quality retrieval flag
1: Medium quality retrieval flag
2: Low quality retrieval flag
3: No retrieval quality flag
}
