% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createRaster.R
\name{goesaodc_createRaster}
\alias{goesaodc_createRaster}
\title{Create a RasterBrick of GOES data}
\usage{
goesaodc_createRaster(
  nc = NULL,
  res = 0.1,
  fun = mean,
  bbox = bbox_CONUS,
  dqfLevel = NULL
)
}
\arguments{
\item{nc}{ncdf4 handle}

\item{res}{resolution of raster in degrees}

\item{fun}{function to use when rasterizing}

\item{bbox}{Bounding box for the region of interest.}

\item{dqfLevel}{Data quality flag level.}
}
\value{
RasterBrick
}
\description{
Create a RasterBrick of GOES AOD data including data points
with the specified resolution and function, and within the specified bounding
box and data quality flag level. 

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or 
\code{raster::extent()}.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadAOD(
  satID = "G17", 
  datetime = "2019-10-27 16", 
  timezone = "UTC",
  verbose = TRUE 
  )

netCDF <- "OR_ABI-L2-AODC-M6_G17_s20193001616196_e20193001618569_c20193001621030.nc"
nc <- goesaodc_openFile(netCDF)

bbox <- c(-129, -110, 29, 50)

rstr <- goesaodc_createRaster(
  nc, 
  res = 0.05, 
  dqfLevel = 3, 
  bbox = bbox
  )
  
west <- c("washington", "oregon", "california", "nevada", "idaho")
maps::map(database = "state", regions = west)   
raster::plot(rstr, "AOD", col=rev(heat.colors(5)))
maps::map(database = "state", regions = west, add = TRUE)
}
}
