% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createRaster.R
\name{goesaodc_createRaster}
\alias{goesaodc_createRaster}
\title{Create a RasterBrick of GOES data}
\usage{
goesaodc_createRaster(nc = NULL, res = 0.1, fun = mean,
  bbox = bbox_CONUS, dqfLevel = NULL)
}
\arguments{
\item{nc}{ncdf4 handle}

\item{res}{resolution of raster in degrees}

\item{fun}{function to use when rasterizing}

\item{bbox}{Bounding box for the region of interest.}

\item{dqfLevel}{Data quality flag level.}
}
\value{
RasterBrick
}
\description{
Create a RasterBrick of GOES AOD data including data points
with the specified resolution and function, and within the specified bounding
box and data quality flag level. 

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or 
\code{raster::extent()}.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadAOD(
  satID = "G16", 
  datetime = "201924918", 
  timezone = "UTC", 
  isJulian = TRUE)

netCDF <- "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc"
 
nc <- goesaodc_openFile(netCDF)
rstr <- goesaodc_createRaster(nc, res = 0.1, dqfLevel = 2)
raster::plot(rstr, "AOD")
maps::map("state", add = TRUE)
}
}
