% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createRaster.R
\name{goesaodc_createRaster}
\alias{goesaodc_createRaster}
\title{Create a RasterBrick of GOES data}
\usage{
goesaodc_createRaster(nc, res = 0.1, fun = mean, bbox = NULL,
  lonLo = NULL, lonHi = NULL, latLo = NULL, latHi = NULL,
  dqfLevel = NULL)
}
\arguments{
\item{nc}{ncdf4 handle}

\item{res}{resolution of raster in degrees}

\item{fun}{function to use when rasterizing}

\item{bbox}{bounding box generated by \code{sp::bbox}. Overrides lonLo, lonHi
latLo, latHi if provided}

\item{lonLo}{lower longitude extent}

\item{lonHi}{upper longitude extent}

\item{latLo}{lower latitude extent}

\item{latHi}{upper latitude extent}

\item{dqfLevel}{data quality flag level}
}
\value{
RasterBrick
}
\description{
Create a RasterBrick of GOES AOD data including data points
with the specified resolution and function, and within the specified extent 
and data quality flag level. Data quality level can take a value of:

0: High quality retrieval flag
1: Medium quality retrieval flag
2: Low quality retrieval flag
3: No retrieval quality flag
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")
netCDF <- system.file("extdata", 
                      "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc", 
                      package = "MazamaSatelliteUtils")
nc <- ncdf4::nc_open(netCDF) 
rstr <- goesaodc_createRaster(nc, res = 0.1, dqfLevel = 2) 
raster::plot(rstr, "AOD")
maps::map("state", add = TRUE)
}
}
