% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createRasterStack.R
\name{goesaodc_createRasterStack}
\alias{goesaodc_createRasterStack}
\title{goesaodc_createRasterStack()}
\usage{
goesaodc_createRasterStack(satID = NULL, datetime = NULL,
  endTime = NULL, var = "AOD", res = 0.1, bbox = c(-125, -65, 24,
  50), dqfLevel = NULL, timezone = "UTC", julian = FALSE)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Desired datetime in any Ymd H [MS] format or \code{POSIXct}.}

\item{endTime}{Desired ending time in any Ymd H [MS] format or \code{POSIXct}}

\item{var}{GOES data variable ("AOD, "DQF" or "ID"). Default: "AOD"}

\item{res}{resolution of raster in degrees. Default: 0.1}

\item{bbox}{Bounding box for raster, Default: c(-125, -65, 24, 50) CONUS}

\item{dqfLevel}{Data quality flag level.}

\item{timezone}{timezone for \code{datetime} and optionally \code{endTime}.}

\item{julian}{Logical specifying if \code{datetime} (and optionally
\code{endTime}) are Julian formatted}
}
\value{
RasterStack
}
\description{
Create a \code{RasterStack} from GOES AOD data files for the
date and hour specified by \code{datetime}. Each \code{RasterLayer} contains
data from one Advanced Baseline Imager (ABI) scan during the specified time
period.

If data for the specified time period is not found in the directory specified
by \code{setSatelliteDataDir()}, it will be downloaded in order to create the
\code{RasterStack}.

The Z axis of the \code{RasterStack} is a character vector where each element
is the time stamp of the scan and has the format YYYYMMDDHHMMSS. This can be
accessed using the \code{raster::getZ()} function. Names of the
\code{RasterStack} are also time stamps of the scan, of the format XHH.MM.SS.

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or
\code{raster::extent()}.
}
\examples{
\dontrun{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

MazamaCoreUtils::initializeLogging(logDir = "~/Data/Logs")

satID <- "G16"
datetime <- "2019-08-12 09:00"
endTime <- "2019-08-12 12:00"
bbox <- c(-124.56624, -116.46350, 41.99179, 46.29203) # Oregon
dqfLevel <- 2
timezone <- "America/Los_Angeles"

rasterStack <- goesaodc_createRasterStack(
satID = satID,
datetime = datetime,
endTime = endTime,
bbox = bbox,
dqfLevel = dqfLevel,
timezone = timezone)

rasterVis::levelplot(rasterStack)
 }
}
