% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createScanRaster.R
\name{goesaodc_createScanRaster}
\alias{goesaodc_createScanRaster}
\title{Create rasters from GOES scans}
\usage{
goesaodc_createScanRaster(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3,
  cellSize = NULL,
  fun = mean
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Data quality flag level; Defaults to NULL.}

\item{cellSize}{Size of Raster cells measured in degrees lon/lat.}

\item{fun}{Function used to summarize multiple point values within a single 
cell; Defaults to \code{mean}.}
}
\description{
Creates a \code{RasterBrick} of AOD and DQF readings from a GOES
scan, or a list of \code{RasterBrick}s from a series of GOES scans.
}
\details{
A single scan can be identified by either giving its 
\code{filename}, or by providing the \code{satID}, \code{datetime}, and 
\code{timezone} (which will be used to determine the closest matching scan). 
A series of scans cannot be specified using filenames. The satellite + time 
info must be given along with an \code{endtime} so that all scans starting at
\code{datetime} up to (but not including) \code{endtime} will be processed.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 46.5)

# Create a raster for a scan specified by satellite + time
goesaodc_createScanRaster(
  satID = "G17",
  datetime = "2020-09-08 17:30",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon,
  cellSize = 0.05
)

# Create a raster for a scan file
goesaodc_createScanRaster(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202530031174_e20202530033547_c20202530035523.nc",
  bbox = bboxOregon,
  dqfLevel = 2,
  cellSize = 0.05
)

# Create a raster from scans averaged over a time range
goesaodc_createScanRaster(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles",
  bbox = bboxOregon,
  cellSize = 0.05
)

# Create a raster for a faulty scan
goesaodc_createScanRaster(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
  bbox = bboxOregon,
  cellSize = 0.05
)
}
}
