% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createScanRaster.R
\name{goesaodc_createScanRaster}
\alias{goesaodc_createScanRaster}
\title{Create Rasters from GOES scans}
\usage{
goesaodc_createScanRaster(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3,
  cellSize = NULL,
  fun = mean
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Data quality flag level; Defaults to NULL.}

\item{cellSize}{Size of Raster cells measured in degrees lon/lat.}

\item{fun}{Function used to summarize multiple point values within a single 
cell; Defaults to \code{mean}.}
}
\description{
Creates a Raster of AOD readings from one or more GOES scans. 
Scans are specified either by filename or by satellite + time information.
}
\details{
If \code{filename} is given, just data for that scan file will be 
processed. Otherwise, scan(s) must be specified by \code{satID}, 
\code{datetime}, and \code{timezone}, which will determine the closest scan 
taken to that time. An average Raster of a series of scans can be requested 
by also supplying \code{endtime}. All scans taken from \code{datetime} 
(inclusive) up to \code{endtime} (exclusive) will be processed by taking the 
average value of each reading point and rasterizing the result (not by 
rasterizing each scan and then taking the average).
}
