% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createScanRaster.R
\name{goesaodc_createScanRaster}
\alias{goesaodc_createScanRaster}
\title{Create a raster from a GOES AOD scan}
\usage{
goesaodc_createScanRaster(
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3,
  cellSize = NULL,
  fun = mean,
  na.rm = FALSE
)
}
\arguments{
\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest. All points outside of 
this area will be removed and therefore won't be rasterized. Defaults to 
CONUS.}

\item{dqfLevel}{Data quality flag level. Defaults to 3.}

\item{cellSize}{Size of Raster cells measured in degrees lon/lat.}

\item{fun}{Function used to summarize point values within a single cell. 
Defaults to \code{mean}.}

\item{na.rm}{Logical flag determining whether to remove \code{NA} values 
before summarizing them with `fun`. Defaults to \code{FALSE}.}
}
\value{
\code{RasterLayer} or \code{RasterBrick} of AOD values.
}
\description{
Creates a \code{RasterLayer} of AOD readings from a GOES scan, 
file or a \code{RasterBrick} of \code{RasterLayer}s from a series of files.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 46.5)

scanFiles <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles"
)

# Create a raster for a scan file
goesaodc_createScanRaster(
  filename = scanFiles[1],
  bbox = bboxOregon,
  dqfLevel = 2,
  cellSize = 0.05
)

# Create a raster from scans averaged over a time range
goesaodc_createScanRaster(
  filename = scanFiles,
  bbox = bboxOregon,
  cellSize = 0.05
)

# Create a raster for a faulty scan
goesaodc_createScanRaster(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
  bbox = bboxOregon,
  cellSize = 0.05
)
}
}
