% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createScanSpdf.R
\name{goesaodc_createScanSpdf}
\alias{goesaodc_createScanSpdf}
\title{Create an SPDF from a GOES AOD scan}
\usage{
goesaodc_createScanSpdf(filename = NULL, bbox = bbox_CONUS, dqfLevel = 3)
}
\arguments{
\item{filename}{Name of a scan file.}

\item{bbox}{Bounding box for the region of interest. All points outside of 
this area will be removed. Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All points with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}
}
\value{
\code{SpatialPointsDataFrame} of AOD values.
}
\description{
Creates a \code{SpatialPointsDataFrame} of AOD readings from a 
GOES scan file, or a list of \code{SpatialPointsDataFrame}s from a series of 
files.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

bboxOregon <- c(-125, -116, 42, 47)

scanFiles <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "America/Los_Angeles"
)

# Create an SPDF from a single scan file
goesaodc_createScanSpdf(
  filename = scanFiles[1],
  bbox = bboxOregon,
  dqfLevel = 2
)

# Create a list of SPDFs from multiple scan files
goesaodc_createScanSpdf(
  filename = scanFiles,
  bbox = bboxOregon
)

# Create an empty SPDF from a faulty scan
goesaodc_createScanSpdf(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
  bbox = bboxOregon
)
}
}
