% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createSingleScanSPDF.R
\name{goesaodc_createSingleScanSPDF}
\alias{goesaodc_createSingleScanSPDF}
\title{Create spatial points for a single scan}
\usage{
goesaodc_createSingleScanSPDF(
  satID = NULL,
  datetime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = NULL
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{The name of the scan file.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Data quality flag level; Defaults to NULL.}
}
\description{
Creates a SpatialPointsDataFrame from a single scan file 
specified either by a filename or through satellite and time information.
}
\details{
If \code{filename} is given, \code{satID}, \code{datetime}, and
\code{timezone} will be ignored.
}
