% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createSpatialPoints.R
\name{goesaodc_createSpatialPoints}
\alias{goesaodc_createSpatialPoints}
\title{Create a SpatialPointsDataFrame of GOES data}
\usage{
goesaodc_createSpatialPoints(nc = NULL, bbox = c(-125, -65, 24, 50),
  dqfLevel = NULL)
}
\arguments{
\item{nc}{ncdf4 handle.}

\item{bbox}{Bounding box for the region of interest.}

\item{dqfLevel}{Data quality flag level.}
}
\value{
SpatialPointsDataFrame
}
\description{
Create a SpatialPointsDataFrame of GOES data including data
within the specified bounding box and data quality flag level. 

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or 
\code{raster::extent()}.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

netCDF <- goesaodc_downloadAOD(
  satID = "G17", 
  datetime = "2019-10-27 10", 
  timezone = "America/Los_Angeles")[1]
  
kincade_bbox <- c(-126, -119, 35, 40)   
  
nc <- goesaodc_openFile(netCDF)
sp <- goesaodc_createSpatialPoints(nc, dqfLevel = 3, bbox = kincade_bbox)
maps::map(database = "state", "regions" = c("california"), xlim = c(-126, -113.5))
goesaodc_plotSpatialPoints(sp, cex = 0.2, add = TRUE)
}
}
