% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createSpatialPoints.R
\name{goesaodc_createSpatialPoints}
\alias{goesaodc_createSpatialPoints}
\title{Create a SpatialPointsDataFrame of GOES data}
\usage{
goesaodc_createSpatialPoints(nc, bbox = NULL, dqfLevel = NULL)
}
\arguments{
\item{nc}{ncdf4 handle.}

\item{bbox}{Bounding box for the region of interest.}

\item{dqfLevel}{Data quality flag level.}
}
\value{
SpatialPointsDataFrame
}
\description{
Create a SpatialPointsDataFrame of GOES data including data
within the specified bounding box and data quality flag level. 

The \code{dqfLevel} parameter can take a value of:

\itemize{
\item{0}{ -- High quality retrieval flag}
\item{1}{ -- Medium quality retrieval flag}
\item{2}{ -- Low quality retrieval flag}
\item{3}{ -- No retrieval quality flag}
}

The \code{bbox} parameter can be a vector of floats in c(lonLo, lonHi, latLo,
latHi) order or the return value from \code{sp::bbox()} or 
\code{raster::extent()}.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadAOD(
  satID = "G16", 
  datetime = "201924918", 
  timezone = "UTC", 
  isJulian = TRUE)

netCDF <- "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc"
nc <- goesaodc_openFile(netCDF)
sp <- goesaodc_createSpatialPoints(nc, dqfLevel = 2)
maps::map("state")
goesaodc_plotSpatialPoints(sp, cex = 0.3, add = TRUE)
}
}
