% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createSpatialPoints.R
\name{goesaodc_createSpatialPoints}
\alias{goesaodc_createSpatialPoints}
\title{Create a SpatialPointsDataFrame of GOES data}
\usage{
goesaodc_createSpatialPoints(nc, bbox = NULL, lonLo = NULL,
  lonHi = NULL, latLo = NULL, latHi = NULL, dqfLevel = NULL)
}
\arguments{
\item{nc}{ncdf4 handle}

\item{bbox}{bounding box generated by \code{sp::bbox}. Overrides lonLo, lonHi
latLo, latHi if provided}

\item{lonLo}{lower longitude extent}

\item{lonHi}{upper longitude extent}

\item{latLo}{lower latitude extent}

\item{latHi}{upper latitude extent}

\item{dqfLevel}{data quality flag level}
}
\value{
SpatialPointsDataFrame
}
\description{
Create a SpatialPointsDataFrame of GOES data including data
within the specified extent and data quality flag level. Data quality level
can take a value of:

0: High quality retrieval flag
1: Medium quality retrieval flag
2: Low quality retrieval flag
3: No retrieval quality flag
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")
netCDF <- system.file("extdata", 
                      "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc", 
                      package = "MazamaSatelliteUtils")
nc <- ncdf4::nc_open(netCDF) 
sp <- goesaodc_createSpatialPoints(nc, dqfLevel = 1)
maps::map("state")
goesaodc_plotSpatialPoints(sp, cex = 0.3, add = TRUE)
}
}
