% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createTibble.R
\name{goesaodc_createTibble}
\alias{goesaodc_createTibble}
\title{Create a tibble from ncdf4 handle}
\usage{
goesaodc_createTibble(nc = NULL, bbox = bbox_CONUS, verbose = FALSE)
}
\arguments{
\item{nc}{ncdf4 handle or a list of handles.}

\item{bbox}{Geographic extent of area of interest; Defaults to CONUS.}

\item{verbose}{Logical flag to increase messages while processing data.}
}
\value{
Tibble (dataframe) with NetCDF variables and associated locations.
}
\description{
Create a tibble with columns: AOD, DQF, lon and lat.
This information is sufficient to plot as points or create a raster object.
}
\examples{
\donttest{
# Tibble based on full extent of Gridfile
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadAOD(
  satID = "G17", 
  datetime = "2019-10-27 10", 
  timezone = "America/Los_Angeles" 
  )

ncFile <- goesaodc_listFiles(
  satID = "G17", 
  datetime = "2019-10-27 10", 
  timezone = "America/Los_Angeles")[1]
                     
nc <- goesaodc_openFile(ncFile)

tbl <- goesaodc_createTibble(nc)
head(tbl)

# Tibble based on BBOX filtered extent of tibble
library(MazamaSatelliteUtils)

ncFile <- goesaodc_listFiles(
  satID = "G17", 
  datetime = "2019-10-27 10", 
  timezone = "America/Los_Angeles")[1]

nc <- goesaodc_openFile(ncFile)

# 2019 Kincade fire in Sonoma county
bbox <- c(-124, -120, 36, 39)

filtered_tbl <- goesaodc_createTibble(nc, bbox)
head(filtered_tbl)

}
}
