% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_createTibble.R
\name{goesaodc_createTibble}
\alias{goesaodc_createTibble}
\title{Create a tibble from ncdf4 handle}
\usage{
goesaodc_createTibble(nc)
}
\arguments{
\item{nc}{ncdf4 handle}
}
\value{
Tibble (dataframe) with NetCDF variables and associated locations.
}
\description{
Create a tibble with columns: AOD, DQF, lon and lat.
This information is sufficient to plot as points or create a raster object.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")

netCDF <- system.file(
  "extdata", 
  "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc", 
  package = "MazamaSatelliteUtils"
)
                     
nc <- goesaodc_openFile(netCDF)

tbl <- goesaodc_createTibble(nc)
head(tbl)
}
}
