% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_downloadAOD.R
\name{goesaodc_downloadAOD}
\alias{goesaodc_downloadAOD}
\title{Download GOES 16 AOD data}
\usage{
goesaodc_downloadAOD(date = NULL, hour = NULL, julianDate = NULL,
  baseUrl = "https://tools-1.airfire.org/Satellite/GOES-16/AODC")
}
\arguments{
\item{date}{desired date (integer, character representing YYYYMMDD or \code{POSIXct})}

\item{hour}{UTC hour for data (HH)}

\item{julianDate}{desired date on Julian calendar (YYYYDDD). Ignored if \code{date} is specified.}

\item{baseUrl}{base URL for data queries}
}
\value{
Vector of downloaded filepaths.
}
\description{
Download all GOES 16 .nc files for the given date and hour to
the directory specified by \code{setSatelliteDataDir()}.
}
\seealso{
\code{\link{setSatelliteDataDir}}
}
