% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_downloadAOD.R
\name{goesaodc_downloadAOD}
\alias{goesaodc_downloadAOD}
\title{Download GOES-16 or GOES-17 AOD data}
\usage{
goesaodc_downloadAOD(satID = NULL, datetime = NULL, endTime = NULL,
  timezone = "UTC", isJulian = FALSE, verbose = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/")
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Desired datetime in any Ymd H [MS] format or \code{POSIXct}}

\item{endTime}{Desired ending time in any Ymd H [MS] format or \code{POSIXct}}

\item{timezone}{Timezone used to interpret \code{datetime} and \code{endTime}}

\item{isJulian}{Logical specifying if \code{datetime} (and optionally 
\code{endTime}) are Julian formatted}

\item{verbose}{Logical parameter specifying whether download status should be
shown.}

\item{baseUrl}{Base URL for data queries.}
}
\value{
Invisibly returns a vector of local files matching the requested \code{datetime}.
}
\description{
Download all GOES 16 or 17 NetCDF files for the given
\code{datetime} to the directory specified by \code{setSatelliteDataDir()}.
If \code{datetime} is specified to the hour and \code{fullDay} is not
explicitly set to TRUE', only files for that hour will be downloaded. If the
optional \code{endTime} is specified, all files that exist for the time 
period between \code{datetime} and \code{endTime} will be downloaded. If 
\code{timezone} is not specified, "UTC" is assumed.

The vector of files returned includes all files in \code{satelliteDataDir}
that match the requested date whether they were download previously or with
this call.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

# Single hour
files <- goesaodc_downloadAOD("G17", "2019-09-06 18") 
print(files)

# Specific time range
files <- goesaodc_downloadAOD(
  satID = "G16",
  datetime = "2019-09-06 08:00",
  endTime = "2019-09-06 12:00",
  timezone = "America/Los_Angeles"
)
print(files)

# Julian timestamp
files <- goesaodc_downloadAOD(
  satID = "G16",
  datetime = "201924915",
  isJulian = TRUE
)
print(files)
}
}
\seealso{
\code{\link{setSatelliteDataDir}}
}
