% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_downloadAOD.R
\name{goesaodc_downloadAOD}
\alias{goesaodc_downloadAOD}
\title{Download GOES AOD data for a specific date}
\usage{
goesaodc_downloadAOD(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  isJulian = FALSE,
  verbose = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/"
)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Datetime in any Ymd H [MS] format or \code{POSIXct}.}

\item{endtime}{End time in any Ymd H [MS] format or \code{POSIXct} 
(exclusive).}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{isJulian}{Logical value determining whether \code{datetime} should be 
interpreted as a Julian date with day of year as a decimal number; Defaults 
to FALSE.}

\item{verbose}{Logical flag to print download progress; Defaults to FALSE.}

\item{baseUrl}{URL of remote database; Defaults to 
"https://tools-1.airfire.org/Satellite/".}
}
\value{
Invisibly returns a vector of local files matching the requested 
\code{datetime}.
}
\description{
Downloads a GOES satellite's NetCDF files for the given 
\code{datetime} to the directory specified by \code{setSatelliteDataDir()}.
If \code{endtime} is specified, all files from \code{startime} up to
(but not including) \code{endtime} will be downloaded. Otherwise, only files 
for the \code{datetime} hour will be retrieved.

The vector of files returned includes all files in \code{satelliteDataDir}
that match the requested date whether they were download previously or with
this call.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

# Single hour
files <- goesaodc_downloadAOD("G17", "2019-09-06 18") 
print(files)

# Time range
files <- goesaodc_downloadAOD(
  satID = "G16",
  datetime = "2019-09-06 08:00",
  endtime = "2019-09-06 12:00",
  timezone = "America/Los_Angeles"
)
print(files)

# Julian timestamp
files <- goesaodc_downloadAOD(
  satID = "G16",
  datetime = "201924915",
  isJulian = TRUE
)
print(files)
}
}
\seealso{
\code{\link{setSatelliteDataDir}}
}
