% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_downloadAOD.R
\name{goesaodc_downloadAOD}
\alias{goesaodc_downloadAOD}
\title{Download GOES-16 or GOES-17 AOD data}
\usage{
goesaodc_downloadAOD(satId = NULL, startdate = NULL, jdate = NULL,
  baseUrl = "https://tools-1.airfire.org/Satellite/", quiet = FALSE)
}
\arguments{
\item{satId}{ID of the source GOES satellite}

\item{startdate}{desired date in any Y-m-d [H] format or \code{POSIXct}}

\item{jdate}{desired date in as a Julian date string, i.e. as seen in the
netcdf filenames}

\item{baseUrl}{base URL for data queries}

\item{quiet}{if TRUE, suppress status messages and progress bar}
}
\value{
Vector of downloaded filepaths.
}
\description{
Download all GOES 16 NetCDF files for the given \code{startdate} 
to the directory specified by \code{setSatelliteDataDir()}. If 
\code{startdate} is specified to the hour, only files for that hour will be 
downloaded. If \code{startdate} is specified only to the day, all files for 
that day will be downloaded.
}
\examples{
\donttest{
setSatelliteDataDir("~/Data/Satellite")

date <- lubridate::ymd_h("2019-05-16 16", tz = "UTC")
files <- goesaodc_downloadAOD("G16", date)
print(files)

date <- 2019051616
files <- goesaodc_downloadAOD("G16", date)
print(files)
}
}
\seealso{
\code{\link{setSatelliteDataDir}}
}
