% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_downloadDaytimeAOD.R
\name{goesaodc_downloadDaytimeAOD}
\alias{goesaodc_downloadDaytimeAOD}
\title{Download GOES AOD data for the daylight hours of a specific date and 
location}
\usage{
goesaodc_downloadDaytimeAOD(
  satID = NULL,
  datetime = NULL,
  longitude = NULL,
  latitude = NULL,
  bbox = NULL,
  timezone = NULL,
  isJulian = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Datetime in any Ymd H [MS] format or \code{POSIXct}.}

\item{longitude}{Longitude of the location of interest in decimal degrees E.}

\item{latitude}{Latitude of the location of interest in decimal degrees N.}

\item{bbox}{Bounding box for the region of interest.}

\item{timezone}{Timezone used to interpret \code{datetime}.}

\item{isJulian}{Logical value determining whether \code{datetime} should be 
interpreted as a Julian date with day of year as a decimal number; Defaults 
to FALSE.}

\item{verbose}{Logical flag to print download progress; Defaults to FALSE.}
}
\value{
Invisibly returns a vector of local files matching the requested 
\code{datetime}.
}
\description{
Downloads a GOES satellite's daylight NetCDF files for the given 
\code{datetime} to the directory specified by \code{setSatelliteDataDir()}.
See \code{getDaylightHours()} for details regarding how daylight hours are 
calculated.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

goesaodc_downloadDaytimeAOD(
  satID = "G16",
  datetime = "2019-09-06", 
  timezone = "America/Los_Angeles"
)
}

}
