% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_getDaylightHours.R
\name{goesaodc_getDaylightHours}
\alias{goesaodc_getDaylightHours}
\title{goesaodc_getDaylightHours}
\usage{
goesaodc_getDaylightHours(datetime = NULL, latitude = NULL,
  longitude = NULL, timezone = NULL, bbox = NULL, julian = FALSE)
}
\arguments{
\item{datetime}{Default: NULL}

\item{latitude}{Default: NULL}

\item{longitude}{Default: NULL}

\item{timezone}{Default: NULL}

\item{julian}{Default: FALSE}
}
\value{
day_info List which contains sunrise and sunset times as POSIXt 
values
}
\description{
Returns the sunrise and sunset hours for a location, timezone, 
or bounding box.
}
\examples{
\dontrun{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

EX 1. Get Sunrise and Sunset by Lat/Lon and Datetime
dayInfo <- goesaodc_getDaylightHours(
datetime = "2019-09-06", 
longitude = -123.245, 
latitude = 42.861)

EX 2. Get Sunrise and Sunset by BBOX and datetime
dayInfo <- goesaodc_getDaylightHours(
datetime = "2019-09-06",
bbox = c(-124.566, -116.463, 41.991, 46.292)) # Oregon

EX 3. Get Sunrise and Sunset by Datetime and Timezone
dayInfo <- goesaodc_getDaylightHours(
datetime = "2019-09-06",
timezone = "America/Los_Angeles")
}
}
