% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_getProjection}
\alias{goesaodc_getProjection}
\title{Get projection metadata from GOES-R ABI ncdf4 object}
\usage{
goesaodc_getProjection(nc = NULL)
}
\arguments{
\item{nc}{ncdf4 handle.}
}
\value{
list of metadata
}
\description{
Return the projection information of `nc`.
}
