% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listDaytimeFiles.R
\name{goesaodc_listDaytimeFiles}
\alias{goesaodc_listDaytimeFiles}
\title{List GOES AOD files that are available for a specific date and 
location during the hours between sunlight and sunset.}
\usage{
goesaodc_listDaytimeFiles(satID = NULL, datetime = NULL,
  longitude = NULL, latitude = NULL, bbox = NULL,
  useRemote = FALSE, timezone = NULL, isJulian = FALSE)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Specific date for which daylight hours are requested.}

\item{longitude}{Longitude of the location of interest in decimal degrees E.}

\item{latitude}{Latitude of the location of interest in decimal degrees N.}

\item{bbox}{Bounding box for the region of interest, Default: CONUS.}

\item{useRemote}{Logical specifying whether to look for files in 
\code{getSatelliteDataDir()} or \code{baseUrl}.}

\item{timezone}{Timezone in which to interpret the \code{datetime}.}

\item{isJulian}{Logical value determining whether datetime should be 
interpreted as a Julian date with day of year as a decimal number.}
}
\value{
Vector of filenames.
}
\description{
List all daylight files available, either locally in directory
specified with \code{setSatelliteDir}, or optionally on remote file server 
by using the \code{useRemote} parameter.  See \code{getDaylightHours()} for 
details regarding how daylight hours are calculated.
}
\examples{
\donttest{

# List locally available files for a specific date by timezone

library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

goesaodc_listDaytimeFiles(satID = "G16",
datetime = "2019-09-06", 
timezone = "America/Los_Angeles")

# List remote files available for specific date by timezone

library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

goesaodc_listDaytimeFiles(satID = "G16", 
datetime = "2019-09-06", 
timezone = "America/Los_Angeles",
useRemote = TRUE)
}


}
