% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listFiles.R
\name{goesaodc_listFiles}
\alias{goesaodc_listFiles}
\title{List downloaded GOES AOD files for a specified date and hour}
\usage{
goesaodc_listFiles(satID = NULL, startdate = NULL, jdate = NULL,
  fullDay = FALSE)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{startdate}{desired date in any Ymd [H] format or \code{POSIXct}}

\item{jdate}{desired date in as a Julian date string, i.e. as seen in the
netcdf filenames}

\item{fullDay}{Specifies that the user wants an entire day's worth of data}
}
\value{
Vector of filenames.
}
\description{
Retrieve a list of GOES AOD files available in the
\code{satelliteDataDir} for a specified date and hour.

Note that all files for a particular hour will be returned even if the
incoming \code{startdate} or \code{jdate} is specified to the minute or
second.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

date_with_hour <- "2019-09-06 16"
goesaodc_listFiles(satID = "G16", startdate = date_with_hour)

jdate <- "201924916"
goesaodc_listFiles(satID = "G17", jdate = jdate, fullDay = TRUE)

day_only <- "2019-09-06"
goesaodc_listFiles(satID = "G16", startdate = day_only)
}
}
