% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listFiles.R
\name{goesaodc_listFiles}
\alias{goesaodc_listFiles}
\title{List available GOES AOD files for a specified date and hour}
\usage{
goesaodc_listFiles(
  satID = NULL,
  datetime = NULL,
  endTime = NULL,
  useRemote = FALSE,
  timezone = "UTC",
  isJulian = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/"
)
}
\arguments{
\item{satID}{ID of the source GOES satellite (G16 or G17).}

\item{datetime}{Desired datetime in any Ymd H [MS] format or \code{POSIXct}}

\item{endTime}{Desired ending time in any Ymd H [MS] format or \code{POSIXct}}

\item{useRemote}{Logical specifying whether to look for files in 
\code{getSatelliteDataDir()} or \code{baseUrl}.}

\item{timezone}{Timezone used to interpret \code{datetime} and \code{endTime}}

\item{isJulian}{Logical specifying if \code{datetime} (and optionally 
\code{endTime}) are Julian formatted}

\item{baseUrl}{Base URL for data queries.}
}
\value{
Vector of filenames.
}
\description{
Retrieve a list of GOES AOD files available in the
\code{satelliteDataDir} or at \code{baseUrl} for a specified date and hour.
}
\note{
All files for a particular hour will be returned even if the
incoming \code{startdate} or \code{jdate} is specified to the minute or
second.  By default, the local directory set by \code{setSatelliteDir()} is 
searched. \code{useRemote = TRUE} will search \code{baseURL} for all files 
that are available within the specified timeframe.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

# Local files for a UTC hour
goesaodc_listFiles("G17", "2019-09-06 18") 

# Remote files for a time range in local time
goesaodc_listFiles(
  satID = "G16",
  datetime = "2019-09-06 06:00",
  endTime = "2019-09-06 18:00",
  timezone = "America/Los_Angeles",
  useRemote = TRUE
)
}
}
