% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listFiles.R
\name{goesaodc_listFiles}
\alias{goesaodc_listFiles}
\title{List downloaded GOES AOD files for a specified date and hour}
\usage{
goesaodc_listFiles(satId = NULL, startdate = NULL, jdate = NULL)
}
\arguments{
\item{satId}{ID number of the source GOES satellite}

\item{startdate}{desired date in any Y-m-d [H] format or \code{POSIXct}}

\item{jdate}{desired date in as a Julian date string, i.e. as seen in the
netcdf filenames}
}
\value{
Vector of filenames.
}
\description{
Retrieve a list of GOES AOD files available in the
\code{satelliteDataDir} for a specified date and hour.

Note that all files for a particular hour will be returned even if the
incoming \code{startdate} or \code{jdate} is specified to the minute or
second.
}
\examples{
\dontrun{
setSatelliteDataDir("~/Data/Satellite")

date <- lubridate::ymd_h("2019-05-16 16", tz = "UTC")
files <- goesaodc_listFiles("G16", date)
print(files)
}
}
