% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_listScanFiles.R
\name{goesaodc_listScanFiles}
\alias{goesaodc_listScanFiles}
\title{List GOES AOD scan files}
\usage{
goesaodc_listScanFiles(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  isJulian = FALSE,
  useRemote = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/"
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime as a Ymd HMS or Julian formatted string, or a 
\code{POSIXct}.}

\item{endtime}{End time as a Ymd HMS or Julian formatted string, or a 
\code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}. Defaults to UTC.}

\item{isJulian}{Logical value determining whether \code{datetime} (and
optionally \code{endtime}) should be interpreted as a Julian date with day of
 year as a decimal number. Defaults to FALSE.}

\item{useRemote}{Logical specifying whether to look for files in 
\code{getSatelliteDataDir()} or \code{baseUrl}. Defaults to FALSE.}

\item{baseUrl}{URL of remote database. Defaults to 
"https://tools-1.airfire.org/Satellite/".}
}
\value{
Vector of scan filenames.
}
\description{
Lists GOES AOD scan files for a given datetime or time range. If
just \code{datetime} is given, then only the scan file closest to that time
will be listed. If \code{endtime} is specified as well, then all scans from
\code{datetime} up to (but not including) \code{endtime} will be listed.
}
\details{
If \code{useRemote=FALSE} and there are no scan files available for 
the day of the requested \code{datetime}, then \code{NULL} will be returned.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:30",
  timezone = "UTC",
  useRemote = TRUE
)

goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2020-09-08 12:00",
  endtime = "2020-09-08 13:00",
  timezone = "UTC",
  useRemote = TRUE
)
}
}
