% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_openFile.R
\name{goesaodc_openFile}
\alias{goesaodc_openFile}
\title{Open a GOES AOD netCDF file}
\usage{
goesaodc_openFile(filename = NULL, dataDir = getSatelliteDataDir())
}
\arguments{
\item{filename}{file name of GOES AOD file}

\item{dataDir}{data directory containing \code{filename}}
}
\value{
nc handle.
}
\description{
Uses \code{ncdf4::nc_open()} to open a NetCDF file and returns
the \code{nc} handle. By default, the directory specified with 
\code{setSatelliteDataDir()} is searched for the file.

Errors generated by `nc_open()` are unhandled and will pass through to the
console or calling function.
}
\examples{
\dontrun{
setSatelliteDataDir("~/Data/Satellite")

netCDF <- system.file(
  "extdata", 
  "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc", 
  package = "MazamaSatelliteUtils"
)

nc <- goesaodc_openFile(netCDF)
}
}
\seealso{
\code{\link{getSatelliteDataDir}}
}
