% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_openFile.R
\name{goesaodc_openFile}
\alias{goesaodc_openFile}
\title{Open a GOES AOD netCDF file}
\usage{
goesaodc_openFile(filename = NULL, dataDir = getSatelliteDataDir())
}
\arguments{
\item{filename}{File name of GOES AOD file.}

\item{dataDir}{Data directory containing \code{filename}; Defaults to 
\code{getSatelliteDataDir()}.}
}
\value{
nc handle
}
\description{
Opens a NetCDF file and returns the \code{nc} handle. By 
default, the directory specified with \code{setSatelliteDataDir()} is 
searched for the file.
}
\details{
Errors generated by `nc_open()` are not handled and will pass through to the
console or calling function.
}
\examples{
\dontrun{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadAOD(
  satID = "G16", 
  datetime = "201924918", 
  timezone = "UTC", 
  isJulian = TRUE
)
  
netCDF <- "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc"

nc <- goesaodc_openFile(netCDF)
}
}
\seealso{
\code{\link{getSatelliteDataDir}}
}
