% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotScanPoints.R
\name{goesaodc_plotScanPoints}
\alias{goesaodc_plotScanPoints}
\title{Create an AOD points plot}
\usage{
goesaodc_plotScanPoints(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3,
  breaks = NULL,
  pointSize = 0.5
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{The name of the scan file.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Data quality flag level; Defaults to 3.}

\item{breaks}{Vector of AOD values to use as palette breaks.}

\item{pointSize}{Size of plot points; Defaults to 0.5.}
}
\description{
Creates a plot of AOD points using GOES scans. If no 
\code{endtime} is given, then only the scan closest to \code{datetime} will
be used. If an \code{endtime} is provided, then the plotted points will be
the average of the scans in the given time range. If \code{filename} is 
given, just the points for that scan file will be drawn.
}
