% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotScanPoints.R
\name{goesaodc_plotScanPoints}
\alias{goesaodc_plotScanPoints}
\title{Plot a GOES AOD scan as points}
\usage{
goesaodc_plotScanPoints(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3,
  pointSize = 0.5,
  pointShape = 15,
  paletteName = "YlOrRd",
  breaks = NULL,
  limits = NULL,
  includeMap = FALSE,
  zoom = NULL,
  stateCodes = NULL,
  title = NULL
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}; Defaults to UTC.}

\item{filename}{Scan file name.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}

\item{pointSize}{Size of plot points; Defaults to 0.5.}

\item{pointShape}{Shape of the plot points (index); Defaults to 15 (filled 
square).}

\item{paletteName}{The name of an RColorBrewer palette; Defaults to 'YlOrRd'.}

\item{breaks}{Vector of AOD values to use as palette breaks.}

\item{limits}{Upper and lower AOD values to use as color scale bounds. 
Setting this guarantees that the color legend is displayed even if the scan 
has nothing but NA AOD values.}

\item{pointAlpha}{Alpha value of the points. Defaults to 0.75.}

\item{includeMap}{Logical flag to draw a topographic map image under the 
raster. Since the image is Mercator projected, the plot coordinate system 
will be Mercator projected to match. Defaults to FALSE.}

\item{zoom}{Zoom level of the topographic map, if it is included. Must be an
integer from 1 to 15.}

\item{stateCodes}{Codes of state outlines to draw.}

\item{title}{Title of the plot.}
}
\description{
Draws a GOES AOD scan as a plot of points. A scan can be 
specified in the following ways:
\itemize{
 \item{Satellite + time info: Set by \code{satID} and \code{datetime}. If no 
   \code{endtime} is given, then only the scan closest to \code{datetime} 
   will be used. If an \code{endtime} is provided, then the points plotted
   will be the average of the all scans in the given time range.}
 \item{File name: Set by \code{filename}.}
}
A file name takes precedence over satellite + time info if both are given.
}
