% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotScanPoints.R
\name{goesaodc_plotScanPoints}
\alias{goesaodc_plotScanPoints}
\title{Plot GOES AOD scan points}
\usage{
goesaodc_plotScanPoints(
  spdf = NULL,
  bbox = bbox_CONUS,
  pointSize = 0.5,
  pointShape = 15,
  pointAlpha = NULL,
  paletteName = "YlOrRd",
  paletteBreaks = NULL,
  legendLimits = NULL,
  includeMap = FALSE,
  zoom = NULL,
  stateCodes = NULL,
  title = NULL
)
}
\arguments{
\item{spdf}{A SpatialPointsDataFrame with AOD and DQF variables.}

\item{bbox}{Bounding box for the region of interest; Defaults to CONUS.}

\item{pointSize}{Size of plot points; Defaults to 0.5.}

\item{pointShape}{Shape of the plot points (index); Defaults to 15 (filled 
square).}

\item{pointAlpha}{Transparency of the points If not explicitly defined, it 
will default to 1.0 when \code{includeMap=FALSE} and 0.75 when
\code{includeMap=TRUE}.}

\item{paletteName}{The name of an RColorBrewer palette; Defaults to 'YlOrRd'.}

\item{paletteBreaks}{Vector of AOD values to use as palette breaks.}

\item{legendLimits}{Upper and lower AOD values for the color legend. Setting 
this guarantees that the legend is displayed even if the scan has nothing but
NA AOD values. All values outside the range will be set to NA.}

\item{includeMap}{Logical flag to draw a topographic map image under the 
raster. Since the image is Mercator projected, the plot coordinate system 
will be Mercator projected to match. Defaults to FALSE.}

\item{zoom}{Zoom level of the topographic map, if it is included. Must be an
integer from 1 to 15.}

\item{stateCodes}{Codes of state outlines to draw.}

\item{title}{Title of the plot.}
}
\description{
Plot GOES AOD scan points
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("NaturalEarthAdm1")

bboxKingcadeFire <- c(-124, -120, 36, 39)

scanPoints <- goesaodc_createScanPoints(
  satID = "G17",
  datetime = "2019-10-27 10:00",
  timezone = "America/Los_Angeles",
  bbox = bboxKingcadeFire
)

faultyScanPoints <- goesaodc_createScanPoints(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
  bbox = bboxKingcadeFire
)

# Plot points for a scan
goesaodc_plotScanPoints(
  spdf = scanPoints,
  bbox = bboxKingcadeFire,
  paletteBreaks = c(-Inf, 0, 1, 2, 3, 4, 5, Inf),
  includeMap = TRUE,
  zoom = 8,
  stateCodes = "CA",
  title = "Kincade fire"
)

# Plot points for a scan filled with NA AOD values
goesaodc_plotScanPoints(
  spdf = faultyScanPoints,
  bbox = bboxKingcadeFire,
  legendLimits = c(-1, 6),
  stateCodes = "CA"
)
}
}
