% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotScanRaster.R
\name{goesaodc_plotScanRaster}
\alias{goesaodc_plotScanRaster}
\title{Plot a GOES scan raster}
\usage{
goesaodc_plotScanRaster(
  raster = NULL,
  bbox = bbox_CONUS,
  fun = mean,
  rasterAlpha = NULL,
  paletteColors = c("#FFFFB2", "#BD0026"),
  naColor = "gray50",
  paletteName = "YlOrRd",
  paletteBreaks = NULL,
  legendLimits = NULL,
  includeMap = FALSE,
  zoom = NULL,
  stateCodes = NULL,
  title = NULL,
  legendTitle = NULL
)
}
\arguments{
\item{raster}{\code{RasterLayer} with an AOD variable.}

\item{bbox}{Bounding box for the visible region. Defaults to CONUS.}

\item{fun}{Function used to summarize multiple point values within a single 
cell. Defaults to \code{mean}.}

\item{rasterAlpha}{Transparency of the raster. If not explicitly defined, it 
will default to 1.0 when \code{includeMap=FALSE} and 0.75 when
\code{includeMap=TRUE}.}

\item{paletteColors}{Vector of colors to use as a gradient for the color 
legend. Will be ignored if \code{breaks} is set. Defaults to 
\code{c("#FFFFB2", "#BD0026")}.}

\item{naColor}{Color to shade cells with NA AOD values.}

\item{paletteName}{The name of an RColorBrewer palette. Defaults to 'YlOrRd'.}

\item{paletteBreaks}{Vector of AOD values to use as palette breaks.}

\item{legendLimits}{Upper and lower AOD values for the fill legend. Setting 
this guarantees that the legend is displayed even if the scan has nothing but
NA AOD values. All values outside the range will be set to NA.}

\item{includeMap}{Logical flag to draw a topographic map image under the 
raster. Since the image is Mercator projected, the plot coordinate system 
will be Mercator projected to match. This significantly slows down the 
drawing of rasters. Defaults to FALSE.}

\item{zoom}{Zoom level of the topographic map, if it is included. Must be an
integer from 1 to 15.}

\item{stateCodes}{Codes of state outlines to draw.}

\item{title}{Title of the plot.}

\item{legendTitle}{Title of the plot legend.}
}
\value{
\code{ggplot} of a \code{RasterLayer}.
}
\description{
Plot a GOES scan raster
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("NaturalEarthAdm1")

bboxKingcadeFire <- c(-124, -120, 36, 39)

scanFile <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2019-10-27 10:00",
  timezone = "America/Los_Angeles"
)

scanRaster <- goesaodc_createScanRaster(
  filename = scanFile,
  bbox = bboxKingcadeFire,
  cellSize = 0.05
)

faultyScanRaster <- goesaodc_createScanRaster(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
  bbox = bboxKingcadeFire,
  cellSize = 0.05
)

# Plot a raster for a scan
goesaodc_plotScanRaster(
  raster = scanRaster,
  bbox = bboxKingcadeFire,
  paletteBreaks = c(-Inf, 0, 1, 2, 3, 4, 5, Inf),
  includeMap = TRUE,
  zoom = 8,
  stateCodes = "CA",
  title = "Kincade fire"
)

# Plot a raster for a scan filled with NA AOD values
goesaodc_plotScanRaster(
  raster = faultyScanRaster,
  bbox = bboxKingcadeFire,
  legendLimits = c(-1, 6),
  stateCodes = "CA"
)
}
}
