% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotScanRaster.R
\name{goesaodc_plotScanRaster}
\alias{goesaodc_plotScanRaster}
\title{Create raster plots of AOD data}
\usage{
goesaodc_plotScanRaster(
  satID = NULL,
  datetime = NULL,
  endtime = NULL,
  timezone = "UTC",
  filename = NULL,
  bbox = bbox_CONUS,
  dqfLevel = 3,
  cellSize = NULL,
  fun = mean,
  paletteName = "YlOrRd",
  breaks = NULL,
  limits = NULL,
  rasterAlpha = NULL,
  includeMap = FALSE,
  zoom = NULL,
  stateCodes = NULL,
  title = NULL
)
}
\arguments{
\item{satID}{ID of the source GOES satellite ('G16' or 'G17').}

\item{datetime}{Datetime in Ymd HMS format or a \code{POSIXct}.}

\item{endtime}{End time in Ymd HMS format or a \code{POSIXct}.}

\item{timezone}{Timezone used to interpret \code{datetime} and 
\code{endtime}. Defaults to UTC.}

\item{filename}{The name of the scan file.}

\item{bbox}{Bounding box for the region of interest. Defaults to CONUS.}

\item{dqfLevel}{Allowed data quality level. All readings with a DQF value
above this level will have their AOD values set to NA. Must be either 0, 1, 
2, or 3, with 0 being the highest quality. Defaults to 3.}

\item{cellSize}{Size of Raster cells measured in degrees lon/lat.}

\item{fun}{Function used to summarize multiple point values within a single 
cell; Defaults to \code{mean}.}

\item{paletteName}{The name of an RColorBrewer palette. Defaults to 'YlOrRd'.}

\item{breaks}{Vector of AOD values to use as palette breaks.}

\item{limits}{Upper and lower AOD values to use as color scale bounds. 
Setting this guarantees that the color legend is displayed even if the scan 
has nothing but NA AOD values.}

\item{rasterAlpha}{Transparency of the raster. If not explicitly defined, it 
will default to 1.0 when \code{includeMap=FALSE} and 0.75 when
\code{includeMap=TRUE}.}

\item{includeMap}{Logical flag to draw a topographic map image under the 
raster. Since the image is Mercator projected, the plot coordinate system 
will be Mercator projected to match. This significantly slows down the 
drawing of rasters. Defaults to FALSE.}

\item{zoom}{Zoom level of the topographic map, if it is included. Must be an
integer from 1 to 15.}

\item{stateCodes}{Codes of state outlines to draw.}

\item{title}{Title of the plot.}
}
\description{
Creates a raster plot of AOD data from a GOES scan. If no 
\code{endtime} is given, then only the scan closest to \code{datetime} will
be used. If an \code{endtime} is provided, then the plotted raster will be
the average of the scans in the given time range. If \code{filename} is 
given, just the raster for that scan file will be drawn.
}
