% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotScanSpdf.R
\name{goesaodc_plotScanSpdf}
\alias{goesaodc_plotScanSpdf}
\title{Plot a GOES scan SPDF}
\usage{
goesaodc_plotScanSpdf(
  spdf = NULL,
  bbox = bbox_CONUS,
  pointSize = 0.5,
  pointShape = 15,
  pointAlpha = NULL,
  paletteColors = c("#FFFFB2", "#BD0026"),
  naColor = "gray50",
  paletteName = "YlOrRd",
  paletteBreaks = NULL,
  legendLimits = NULL,
  includeMap = FALSE,
  zoom = NULL,
  stateCodes = NULL,
  title = NULL,
  legendTitle = NULL
)
}
\arguments{
\item{spdf}{\code{SpatialPointsDataFrame} with an AOD variable.}

\item{bbox}{Bounding box for the visible region. Defaults to CONUS.}

\item{pointSize}{Size of plot points. Defaults to 0.5.}

\item{pointShape}{Shape of the plot points (index). Defaults to 15 (filled 
square).}

\item{pointAlpha}{Transparency of the points. If not explicitly defined, it 
will default to 1.0 when \code{includeMap=FALSE} and 0.75 when
\code{includeMap=TRUE}.}

\item{paletteColors}{Vector of colors defining the gradient for the color 
legend. Will be ignored if \code{breaks} is set. Defaults to 
\code{c("#FFFFB2", "#BD0026")}.}

\item{naColor}{Color to shade cells with NA AOD values.}

\item{paletteName}{Name of an RColorBrewer palette. Defaults to 'YlOrRd'.}

\item{paletteBreaks}{Vector of AOD values to use as palette breaks.}

\item{legendLimits}{Upper and lower values for the color legend. Setting this
guarantees that the legend will appear even if the SPDF has nothing but
\code{NA} values. All values outside this range will be set to \code{NA}.}

\item{includeMap}{Logical flag determining whether to draw a topographic map 
image under the points. Since the image is Mercator projected, the plot 
coordinate system will be Mercator projected to match. Defaults to FALSE.}

\item{zoom}{Zoom level of the topographic map, if it is included. Must be an
integer from 1 to 15.}

\item{stateCodes}{Codes of state outlines to draw.}

\item{title}{Title of the plot.}

\item{legendTitle}{Title of the plot legend.}
}
\value{
\code{ggplot} of a \code{SpatialPointsDataFrame}.
}
\description{
Plot a GOES scan SPDF
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
library(MazamaSpatialUtils)

setSatelliteDataDir("~/Data/Satellite")
setSpatialDataDir("~/Data/Spatial")

loadSpatialData("NaturalEarthAdm1")

bboxKingcadeFire <- c(-124, -120, 36, 39)

scanFile <- goesaodc_listScanFiles(
  satID = "G17",
  datetime = "2019-10-27 10:00",
  timezone = "America/Los_Angeles"
)

scanSpdf <- goesaodc_createScanSpdf(
  filename = scanFile,
  bbox = bboxKingcadeFire
)

emptyScanSpdf <- goesaodc_createScanSpdf(
  filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
  bbox = bboxKingcadeFire
)

# Plot SPDF for a scan
goesaodc_plotScanSpdf(
  spdf = scanSpdf,
  bbox = bboxKingcadeFire,
  paletteBreaks = c(-Inf, 0, 1, 2, 3, 4, 5, Inf),
  includeMap = TRUE,
  zoom = 8,
  stateCodes = "CA",
  title = "Kincade fire"
)

# Plot SPDF for a scan filled with NA AOD values
goesaodc_plotScanSpdf(
  spdf = emptyScanSpdf,
  bbox = bboxKingcadeFire,
  legendLimits = c(-1, 6),
  stateCodes = "CA"
)
}
}
