% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_plotSpatialPoints}
\alias{goesaodc_plotSpatialPoints}
\title{Create a quick plot of a GOES AOD SpatialPointsDataFrame}
\usage{
goesaodc_plotSpatialPoints(spatialPoints, var = "AOD", n = 1e+05,
  breaks = NULL, pch = 15, cex = 0.5, paletteName = "YlOrRd",
  add = FALSE)
}
\arguments{
\item{spatialPoints}{SpatialPointsDataFrame}

\item{var}{Variable to plot}

\item{n}{Sample size}

\item{breaks}{vector of color breaks}

\item{pch}{plot character}

\item{cex}{plot symbol scale factor}

\item{paletteName}{RColorBrewer palette name}

\item{add}{logical to add points to current plot}
}
\description{
Quickly subsample and plot points in a GOES AOD 
spatialPointsDataFrame
}
