% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goesaodc_plotSpatialPoints.R
\name{goesaodc_plotSpatialPoints}
\alias{goesaodc_plotSpatialPoints}
\title{Create a quick plot of a GOES AOD SpatialPointsDataFrame}
\usage{
goesaodc_plotSpatialPoints(spatialPoints, var = "AOD", n = 1e+05,
  colBins = 5, breaks = NULL, paletteName = "YlOrRd", pch = 15,
  cex = 0.5, add = FALSE)
}
\arguments{
\item{spatialPoints}{SpatialPointsDataFrame}

\item{var}{Variable to plot}

\item{n}{Sample size}

\item{colBins}{number of color bins}

\item{breaks}{vector of color breaks}

\item{paletteName}{RColorBrewer palette name}

\item{pch}{plot character}

\item{cex}{plot symbol scale factor}

\item{add}{logical specifying whether to create a new plot or add to existing}
}
\description{
Quickly subsample and plot points in a GOES AOD 
spatialPointsDataFrame
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadAOD(
  satID = "G17", 
  datetime = "2019-10-27 10", 
  timezone = "America/Los_Angeles" 
  )
  
kincade_bbox <- c(-126, -119, 35, 40)   

netCDF <- goesaodc_listFiles(
  satID = "G17", 
  datetime = "2019-10-27 10", 
  timezone = "America/Los_Angeles")[1]
  
nc <- goesaodc_openFile(netCDF)
sp <- goesaodc_createSpatialPoints(nc, dqfLevel = 3, bbox = kincade_bbox)
maps::map(database = "state", "regions" = c("california"), xlim = c(-126, -113.5))
goesaodc_plotSpatialPoints(sp, cex = 0.2, add = TRUE)
}
}
