% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goesaodc.R
\name{goesaodc_scaleAOD}
\alias{goesaodc_scaleAOD}
\title{Converts raw, signed short AOD values from GOES 16 & 17 NetCDF files}
\usage{
goesaodc_scaleAOD(aod_data, conversion_factors)
}
\arguments{
\item{aod_data}{raw, signed short AOD data from GOES .nc file}

\item{conversion_factors}{AOD:'_FillValue', AOD:scale_factor, AOD:add_offset
from AOD variable metadata in GOES .nc file}
}
\value{
The scan start time.
}
\description{
Performs a series of conversions to raw AOD values from .nc file
#' \itemize{
\item{0}{ -- Converts '_FillValue' entries to NA}
\item{1}{ -- Converts signed short values into unsigned short values}
\item{2}{ -- Applies AOD:scale_factor and AOD:add_offset to create correctly
scaled AOD values}
}
}
