% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installGoesGrids.R
\name{installGoesGrids}
\alias{installGoesGrids}
\title{Installs GOES grids in satelliteDataDir}
\usage{
installGoesGrids()
}
\description{
Creates data files with geolocation information for GOES-16 
(East) and GOES-17 (West) satellite products.  Checks for the presence of
GOES 16 & GOES 17 .nc (NetCDF) files in directory previously set by 
\code{setSatelliteDataDir}. If present, will read projection and 
coordinate grid information from them and create GOES East and West grids in 
that directory.  If .nc files are missing, will download appropriate ones 
and use them to create the grids.
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")
installGoesGrids()
} 
}
\seealso{
\code{\link{createGoesGrid}}
}
