% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installGoesGrids.R
\name{installGoesGrids}
\alias{installGoesGrids}
\title{Create GOES grids in satelliteDataDir}
\usage{
installGoesGrids()
}
\description{
Creates data files with geolocation information for GOES-16 
(East) and GOES-17 (West) satellite products.  Checks for the presence of
GOES 16 & GOES 17 .nc (NetCDF) files in /inst/extdata. If present, it
will read projection and coordinate grid information from them and create
GOES East and West grids in the directory previously set with 
\code{setSatelliteDataDir()}.

JON: What do you want to do if the .nc files DON'T exist?
TODO: Function currently generates warinings:
"In sqrt(b^2 - 4 * a * c) : NaNs produced"
Check that we can safely ignore them, or even suppressWarningMessages({ ... })
}
\examples{
\donttest{
library(MazamaSatelliteUtils)

setSatelliteDataDir("~/Data/Satellite")
installGoesGrids()
} 
}
