% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{isDaylight}
\alias{isDaylight}
\title{check if specified \code{datetime} is during the daylight hours of 
the specified \code{region}}
\usage{
isDaylight(datetime = NULL, bbox = NULL, timezone = "UTC")
}
\arguments{
\item{datetime}{desired datetime in any Y-m-d H [MS] format or \code{POSIXct}}

\item{bbox}{bounding box for region of interest}

\item{timezone}{timezone in which to interpret the \code{datetime}}
}
\value{
logical
}
\description{
check if specified \code{datetime} is during the daylight hours of 
the specified \code{region}
}
\examples{
\donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

# BBOX AS CREATED BY SP

library(MazamaSpatialUtils)

mx <- subset(SimpleCountries, countryCode == "MX")

mx_bbox <- bbox(mx)

isDaylight(
  datetime = "2019-09-06 12",
  bbox = mx_bbox,
  timezone = "UTC") 

# BBOX AS DEFINED BY CALLING ON raster::extent
G16_filepath <- system.file(
  "extdata", 
  "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc", 
  package = "MazamaSatelliteUtils")

nc <- ncdf4::nc_open(G16_filepath)

raster <- goesaodc_createRaster(nc, res = 0.1, dqfLevel = 2)

extent <- raster::extent(raster)

isDaylight(
  datetime = "2019-09-06 12",
  bbox = extent,
  timezone = "UTC")
}
}
