% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{isDaylight}
\alias{isDaylight}
\title{check if specified \code{datetime} is during the daylight hours of the
specified \code{region}}
\usage{
isDaylight(datetime = NULL, bbox = NULL, timezone = "UTC")
}
\arguments{
\item{datetime}{desired datetime in any Y-m-d H [MS] format or \code{POSIXct}}

\item{bbox}{bounding box for region of interest}

\item{timezone}{timezone in which to interpret the \code{datetime}}
}
\value{
logical
}
\description{
check if specified \code{datetime} is during the daylight hours of the
specified \code{region}
}
